[UBSAN] Update exceptions
Summary:
This diff adds a new sanitizer exception entry to avoid a `vptr` issue
caused by the boost test case timeout in the avalanche unit test
`poll_inflight_timeout`. For some unknown reason this issue only
triggers when building with cmake, and only with clang 7 (8 and 9 are
OK).

Test Plan:
  mkdir -p buildcmake && cd buildcmake
  cmake -GNinja .. -DCMAKE_C_COMPILER=clang-7 -DCMAKE_CXX_COMPILER=clang++-7 -DCMAKE_BUILD_TYPE=Debug -DENABLE_SANITIZERS=undefined
  ninja
  export UBSAN_OPTIONS="suppressions=../sanitizer_suppressions/ubsan:print_stacktrace=1:halt_on_error=1"
  ninja check

Reviewers: #bitcoin_abc, deadalnix, jasonbcox

Reviewed By: #bitcoin_abc, jasonbcox

Subscribers: jasonbcox

Differential Revision: https://reviews.bitcoinabc.org/D4708