stop rewinding post-segwit blocks on startup
Summary:
This seems to have been missed in the original removal of segwit.

Introduced in the segwit PR8149, RewindBlockIndex carefully takes care
that when we upgrade to segwit, we erase post-ugprade blocks which could
be missing witness data.

Since segwit doesn't activate ever, this function does nothing but slow
down startup negligibly, by unnecessarily flushing the state to disk. It's
also a lot of untested code which means if we ever need it for something
else, the code is likely to be in a broken state.

Test Plan:
    make check
    test_runner.py

Reviewers: fpelliccioni, Fabien, #bitcoin_abc, deadalnix

Reviewed By: fpelliccioni, #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D4728