[AUTOTOOLS] Don't build the seeder tests if --disable-tests is set
Summary:
The seeder tests are conditioned to the seeder build, but it must be
built only if the seeder is built *AND* the tests are enabled.
If the tests are not enabled there is no check for the boost unit tests
library.

Test Plan:
  ./autogen.sh
  mkdir build && cd build
  ../configure
  make
Check that `bitcoin-seeder` and `test_bitcoin-seeder` are built.

  ../configure --disable-tests
  make
Check that `bitcoin-seeder` is built but not `test_bitcoin-seeder`.

  ../configure --without-seeder
  make
Check that neither `bitcoin-seeder` nor `test_bitcoin-seeder` is built.

Reviewers: #bitcoin_abc, deadalnix, jasonbcox, nakihito

Reviewed By: #bitcoin_abc, jasonbcox, nakihito

Differential Revision: https://reviews.bitcoinabc.org/D4767