Add a facility to extract libsecp256k1 from the repository
Summary: There is a need to have libsecp256k1 on its own for 3rd parties to use. This script can extract libsecp256k1 from a clone of the Bitcoin ABC repository.

Test Plan:
Run the script and ensure the generated repository looks good.

Running the script takes a long time, so plan accordingly.

Reviewers: #bitcoin_abc, jasonbcox

Reviewed By: #bitcoin_abc, jasonbcox

Subscribers: Fabien, Mengerian, jasonbcox

Differential Revision: https://reviews.bitcoinabc.org/D4826