Revert "[mining] Rename several CBlockTemplateEntry members for clarity"
Summary:
This reverts D2862 (commit 714f2fafc7f3b78db90d051f22dc9103f408cebd).
It seems it was part of a large stack that never went anywhere. Besides
that, the title was wrong as this was not rename-only: it added a
variable and changed type of sigops count from integer to unsigned.

This is a bit dangerous since negative sigops counts actually do get
passed in sometimes (as in one of the touched code lines), which means
they are converted to giant sigops count. This didn't actually cause
a bug but seems like an unintended effect for a rename-only diff.

Test Plan: `ninja check`

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D4831