save ScriptExecutionMetrics during CScriptCheck
Summary:
CScriptCheck represents a full closure of the call to VerifyScript.
Currently it just saves the outputted error code; update it to also save
the metrics return value.

A small test (taking advantage of an existing test that uses CScriptCheck) is
included.

Depends on D4918

Test Plan: `ninja all check`

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D4920