[CI] Split build-default into build-diff and build-master
Summary:
The same `build-default` configuration is used for 2 different purposes,
with a branch detection mechanism to select what to run.
This diff makes it 2 different configurations instead, making it easier
to follow and easier to maintain.

Apart from the split itself, there is no change in behavior and coverage
remains the same.

Test Plan:
  ABC_BUILD_NAME=build-diff ./contrib/teamcity/build-configurations.sh
  ABC_BUILD_NAME=build-master ./contrib/teamcity/build-configurations.sh

Reviewers: #bitcoin_abc, deadalnix, jasonbcox

Reviewed By: #bitcoin_abc, jasonbcox

Subscribers: jasonbcox

Differential Revision: https://reviews.bitcoinabc.org/D4925