Move the JUnit file to the temporary directory
Summary:
The `test_runner.py` script generates a JUnit XML ouput file which is
used by the CI to offer comprehensive test report.
The default for it is a constant file name located in the build
directory. The result is that subsequent calls to `test_runner.py` will
override the previous reports if an explicit file name is not passed.

This diff keeps the file name the same but moves it to the temporary
directory, which is unique to the runner. It avoids collisions, makes it
easier to find the report and makes it more obvious which runner it
relates to.

Test Plan:
  # The test doesn't matter, but feature_help runs fast
  ./test/functional/test_runner.py feature_help --nocleanup
Ensure the `junit_result.xml` file is located under the temporary
directory of the runner.

  ./test/functional/test_runner.py -J junit_report.xml --nocleanup feature_help
Ensure the `junit_report.xml` file is located under the temporary
directory of the runner.

  mkdir -p /tmp/junit
  ./test/functional/test_runner.py -J /tmp/junit/report.xml --nocleanup feature_help
Ensure the `/tmp/junit/report.xml` file is created and there is no other
report in the temporary directory.

Reviewers: #bitcoin_abc, deadalnix, jasonbcox

Reviewed By: #bitcoin_abc, jasonbcox

Subscribers: jasonbcox

Differential Revision: https://reviews.bitcoinabc.org/D4938