add a flag that (if unset) zeroes sigchecks reported by VerifyScript
Summary:
We want all the sigchecks counting/caching machinery to be functioning and in
place before the upgrade activation, however we don't actually want sigchecks
to be nonzero until then:
- prior to upgrade the sigchecks total could be too large for the cache, and
  we definitely want high-sigchecks txns to be cached.
- it's easier to have the block sigchecks limit fully in place, but then
  have nothing counting towards it

Effectively, we will be able to simply flip on this flag to cause the
activation of sigchecks limits in consensus and in block template creation.

This flag can be removed (always-on) as soon as the upgrade is past and we
impose a retroactive consensus sigchecks limit.

Depends on D4920

Test Plan: `ninja check`

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Maniphest Tasks: T704

Differential Revision: https://reviews.bitcoinabc.org/D4940