Temporary fix for recent build flakiness
Summary:
There is a path collision encountered in util_tests.cpp::test_LockDirectory when multiple instances of this
test are run in parallel (the tests attempt to work on the same dir and eventually fail a lock check).  This recently
arose when we started running upgraded and non-upgraded unit tests in parallel on CI, thereby running util_tests twice.

There are a few backports that we are missing that may improve the situation, but some quick testing on Core master
indicates that it may not be entirely fixed.

In the mean time, this patch should fix the flakiness associated with this test while I drill down into the problem
for a proper fix.

Test Plan: Run build-diff and build-master on CI many times.

Reviewers: #bitcoin_abc, deadalnix, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D5003