CreateNewBlock: insert entries into block slightly earlier so that correct size is logged
Summary:
I noticed the log message was always printing "CreateNewBlock(): total size: 81 ...",
which is a bit silly.

This moves the pblock population back to where it belongs, right after the
`entries` are determined. To make this work, the coinbase tx is inserted
into pblock directly which is also what we want (and this goes back to what
Core does).

Test Plan: `ninja check-all`

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D5019