[CMAKE] Bump minimum version to 3.13
Summary:
This version brings the `add_link_options` function that will allow to
attach linker flags to targets instead of overriding the
`CMAKE_EXE_LINKER_FLAGS`. It also fixes a bug with empty lists (see
summary in D5033 for details).

There is no major distribution packaging cmake 3.12 and not 3.13 so
there is no impact for developers and package maintainers (Fedora 29 is
an exception but has an update package to 3.14, see
https://pkgs.org/download/cmake).

This time I updated the required version for all the libs as well. Not
doing so makes the build brittle as these libs are using our top level
cmake modules. The version requirement for the `src/CMakeLists.txt` is
removed because it is a duplicate of the top-level `CMakeLists.txt`
file.

Depends on D5033.

Test Plan:
  ninja check

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D5035