[DOC] Add CMake and Ninja to the dependency list
Summary:
This will make it easier for users to determine the minimum required
version. Since there is no strong requirement for Ninja, I used the
version from Xenial which is the oldest one that I know is working, but
previous might work just as well.

Test Plan: Read the doc.

Reviewers: #bitcoin_abc, deadalnix, jasonbcox

Reviewed By: #bitcoin_abc, jasonbcox

Differential Revision: https://reviews.bitcoinabc.org/D5038