split feature_block test into sigops and non-sigops parts
Summary:
feature_block.py tests a lot of quirky and fun cases of sigops
counting and sigops limiting. Since sigops is going to be
removed, it makes sense to separate out the tests so that
feature_block can keep running independent of upgrade, and
the sigops tests can be specialized to run depending on
upgrade.

I've left placeholder blocks in place of the sigops tests.
This is because feature_block.py relies on the chain being
built and blocks being matured in a particular order, and
if we outright skip some blocks we may get unmatured
coinbase problems, perhaps not now but definitely possible
in future if some more things are added / backported into
this test.

Test Plan: `ninja check-functional-extended`

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D5060