move sigops counting from CheckBlock to ContextualCheckBlock
Summary:
Upgrade features can't live in CheckBlock (which doesn't know anything
about the ancestors of the block), and have to live in ContextualCheckBlock.
So in order to deactivate sigops limits, it has to move.

(while we're at it, this simplifies a rather weird loop)

Test Plan: `ninja check-extended`

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Maniphest Tasks: T704

Differential Revision: https://reviews.bitcoinabc.org/D5061