[SECP256K1] CMake: add an install target
Summary:
Add an install target for the secp256k1 library. The gitian descriptors
are updated to not include it as part of the bitcoin-abc gitian build.

Depends on D5098 and D5099.

Test Plan:
  mkdir install
  cmake -GNinja .. -DCMAKE_INSTALL_PREFIX=install
  ninja
  cmake -DCOMPONENT=secp256k1 -P cmake_install.cmake
Check the secp256k1 libraries and headers are installed.

Run the Travis build (see
https://travis-ci.org/Fabcien/secp256k1/builds/643458709).

Run the Gitian builds and check the content of the archives is correct.

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D5100