Automatically add missing braces
Summary:
This diff adds a check to `clang-tidy` for adding the braces automatically.
Configuration files are added to the libraries (leveldb, secp256k1,
univalue) to avoid formatting them.

Depends on D5138.

Test Plan:
Add some missing brace somewhere in a source file if D5120 is not landed yet.
  cmake -GNinja ..
  ninja
Check that the files are updated with braces at build time.

Reset the file to have missing braces, then:
  cmake -GNinja .. -DENABLE_CLANG_TIDY=OFF
  ninja
Check that `clang-tidy` is not run.

Run the Gitian builds.

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Subscribers: deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D5119