add missing swap to CScriptCheck
Summary:
My bad, I missed adding this. It doesn't really matter since
the results of CScriptCheck (neither metrics nor error) are
never used in the parallel setting. In future I'd really like
to make error/metrics not be members of this class anyway,
instead to be optional outparams of the `operator()`.

Test Plan: `ninja check`

Reviewers: #bitcoin_abc, jasonbcox

Reviewed By: #bitcoin_abc, jasonbcox

Differential Revision: https://reviews.bitcoinabc.org/D5225