remove SCRIPT_VERIFY_COMPRESSED_PUBKEYTYPE (aka WITNESS_PUBKEYTYPE)
Summary:
This flag is unused and poorly maintained (we lack all of the tests
that Core has about this, since they used segwit). It also seems to
cause confusion among everyone who sees it for the first time, as nobody
expects we have a flag that we never use.

Enabling this flag at any point in future would render millions of BCH
unspendable, unless we start introducing separate script version
systems (like segwit), in which case reverting this diff will be a very
tiny part of the overall complexity. Even then, it's not clear that we
would borrow the same script interpreter subsystem instead of making a
whole new script.

And in fact there's nothing wrong with uncompressed keys. Even though
taking more space, they're faster to verify, especially in batch
verification where key decompression could perhaps take as much as 20%
of the CPU time when doing large batches.

Test Plan: `ninja check-all`

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D5236