Fix a race condition with rpc ports in check-seeds
Summary:
The RPC port from the first run of test-seeds is not guaranteed to be closed/cleaned up
by the OS by the time the second run on test-seeds occurs.  Without this patch, check-seeds is flaky
and fails at random.

Test Plan:
```
ABC_BUILD_NAME=check-seeds ./build-configuration.sh
```
Run this a few times on CI too

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D5268