Don't use std::quick_exit() as it is poorly supported
Summary:
Despite being standard, it appears that `quick_exit()` is not available
on most non glibc platforms. In particular it breaks the tests build on
OSX.
This diff reverts D5254 in favor of D5232 (credits to @markblundeberg).
Despite being more hacky, it is more portable in definitive.

Test Plan:
On Linux and OSX:
  ninja check
Note that windows is unaffected as the test is not built for this
platform.

  valgrind ./src/test/test_bitcoin --logger=JUNIT  -t util_tests

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Subscribers: markblundeberg

Differential Revision: https://reviews.bitcoinabc.org/D5318