Fix for issue #22, ABC bug whereby windows uninstaller misses some .lnk files
This commit fixes issue #22, whereby the Windows unininstaller was not
correctly removing the "Startup" *.lnk items created by the GUI at runtime
when the user enabled "run this program at startup".

Note that this very minor bug has existed in Bitcoin ABC since 2017, and
we are just now fixing it here in this codebase.