[CMAKE] Fix the check-bitcoin-* targets when running with Xcode
Summary:
This diff makes the unit test runner script use the full path to run
`test_bitcoin`, making it independent of the generator configuration for
Xcode and Visual Studio (see D5338). This applies as well to
`test_bitcoin-qt` and `test_bitcoin-seeder`.

Note: the `check` target still fails and will require more changes to
make it work. This is because the python scripts `bitcoin-util-test.py`
use an hardcoded path to the `bitcoin-tx` binary (relative to the build
directory).

Test Plan:
  cmake -GNinja ..
  ninja check

  cmake -G"Unix Makefiles" ..
  make check

On OSX:
  cmake -GXcode ..
  cmake --build . --config RelWithDebInfo --target check-bitcoin
  cmake --build . --config RelWithDebInfo --target check-bitcoin-qt
  cmake --build . --config RelWithDebInfo --target check-bitcoin-seeder

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Subscribers: deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D5342