Reorganise main window title (#38)
Removes the excessive dash, which was intended to separate network name
(empty string on mainnet) from rest of title, not for a version number.

Note:
* Main window title is always same as exit window title.
* Splash window title has no problems.

Closes #38

Tested as follows:

1. `src/qt/bitcoind`, title is "Bitcoin Cash Node"
2. `src/qt/bitcoind -regtest`, title is "Bitcoin Cash Node [regtest]"

If wallet has a non-empty display name (no idea what that is or how to
get that), it should be appended like "Bitcoin Cash Node [regtest] -
WalletDisplayName".