[SECP256K1] Enable the OpenSSL tests (and benchmark)
Summary:
This diff enables building the OpenSSL tests by default with CMake. In
order to get rid of the dependency, the `SECP256K1_BUILD_OPENSSL_TESTS`
variable can be set to `OFF`. It will also be disabled if the tests are
disabled.

Test Plan:
  cmake -GNinja .. -DSECP256K1_BUILD_TEST=OFF
  grep SECP256K1_BUILD_OPENSSL_TESTS CMakeCache.txt
Check that `SECP256K1_BUILD_OPENSSL_TESTS` is not defined in the cache.

  cmake -GNinja .. -DSECP256K1_BUILD_OPENSSL_TESTS=OFF
  grep SECP256K1_BUILD_OPENSSL_TESTS CMakeCache.txt
Check that `SECP256K1_BUILD_OPENSSL_TESTS` is disabled in the cache.

  cmake -GNinja ..
  grep SECP256K1_BUILD_OPENSSL_TESTS CMakeCache.txt
Check that `SECP256K1_BUILD_OPENSSL_TESTS` is enabled in the cache.
  ninja check-secp256k1
  ninja bench-secp256k1
Check in the output that the benchmark `ecdsa_verify_openssl` is run.

Run the Travis build (see
https://travis-ci.org/Fabcien/secp256k1/builds/657808023).

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Subscribers: deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D5429