Remove BIP9 and BIP145 references from getblocktemplate RPC help
The RPC help for `getblocktemplate` refers to four BIPs (22/23/9/145),
two of which are problematic:

1. It refers to the BIP9 specification, despite none of the
`getblocktemplate` section of BIP9 currently being implemented (this
caused the IFP voting problems in ABC 0.21.0).
2. It refers to the BIP145 specification, but the only visible remnant
of this SegWit specification is that we have a `txid` field mirroring
the `hash` field - referencing this specification thus adds more
confusion than clarity.

This commit removes specification references to these two BIPs.

Tested as follows:

1. Review the diff of this MR.
2. In RPC console, `help getblocktemplate`.
3. Verify the changes.