Fix FindBerkeleyDB suffix paths
Summary:
The FindBerkeleyDB module builds a combination of component and version
to search for the associated library, but use the same combination for
the suffix paths. However the suffix paths are always based on the name
`db` (not `db_cxx`) and can cause the search to fail on some systems.

Test Plan:
  cmake -GNinja ..
  ninja

Run the Gitian builds.

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D5446