[backport] Qt: Replace deprecated QSignalMapper with lambda slots
This is taken directly from Core #16706 (ABC doesn't have this backport
yet).

This fixes deprecated warnings if compiling on Qt >= 5.10.

QSignalMapper was deprecated between Qt 5.10 and 5.13 (and then
undeprecated in 5.14).  QSignalMapper isn't as useful as it once was
because now that C++ has lambdas, it actually doesn't really need to
exist.

At any rate -- this commit is verbatim adapted from Core's #16706.

They completely remove the usage of QSignalMapper in favor of lambdas
that capture some local state.

Motivation: It annoyed me that there were deprecated warning messages
when building BitoinQt.

---

Test Plan:

The signal mapper was used in the following actions in the UI:

  - Right-click on a peer in debug window -> ban peer action
  - Right click on a tx in the transaction tab, view on external block
explorer (requires you configure an "external link" in the preferences).

How to test:

  1. Build bitcoin-qt. Optional: If you want, use Qt 5.10 or above, in which
     case you should see *no* deprecated warnings related to QSignalMapper
  2. Navigate around the UI to the Debug Window and try to ban a peer for
     1h or 1d, etc. It should work immediately.
  3. Go to Preferences -> Display -> Third party transacion URLs and
     specify a URL.  E.g. on testnet: https://explorer.bitcoin.com/tbch/tx/%s
  4. Make sure you have a tx in history, right click a tx, open URL in
     external browser (should be the last option in the context menu).
  5. The action should still work and the URL should open in browser.