Qt: Set AA_EnableHighDpiScaling before QApplication instance creation
This fixes a warning about incorrect usage of Qt at app startup.
The warning in question is:

  Attribute Qt::AA_EnableHighDpiScaling must be set before QCoreApplication
  is created.

Background:

From Qt docs for AA_EnableHighDpiScaling:
https://doc.qt.io/qt-5/qt.html#ApplicationAttribute-enum

"This attribute must be set before QGuiApplication is constructed."

Setting this value before the QApplication is constructed also means it
actually has an effect (which is to use sharter graphics on platforms
that support it).

This has been tested and works 100%.  Note: We do this also in Electron
Cash and I have experience with Qt and its various quirks.

Test plan:

Try running on a relatively recent Qt such as Qt 5.13 without this commit
and observe the console may display the warning.

Next, recompile with this commit.  You should notice the warning disappeared.