Use a sane default version for PPA releases
Summary:
The typical mode of operation is for the version number to be exactly the same as the binary.
There's no reason to not use this, since its readily available. A --version option is provided for
testing and hotfix purposes.  This patch will reduce user error, especially when uploading to production.

Test Plan:
```
./debian-packages.sh --dry-run "Jason B. Cox <jasonbcox@bitcoinabc.org>"
./debian-packages.sh --dry-run --version 0.21.2.1 "Jason B. Cox <jasonbcox@bitcoinabc.org>"
```

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D5482