UniValue performance speedups for .write()
As discussed in issue #51, the UniValue library could use some
improvement. It inefficiently concatenates strings as it serializes
JSON, which costs some extra CPU cycles in extra copies and extra allocations.

As a first pass, this merge request went through the `univalue_write.cpp`
file and made the `UniValue::write()` function internally use a
`std::stringstream` for serializing JSON to string.

This internally uses the same buffer with amortized constant time
insertions at the end, which it writes to (similar to the complexity of
using a single `std::string::append()` or `std::vector::push_back()`).

In my own benchmarks I found roughly a 5% performance improvement when
serializing.

The main savings in this commit probably comes from not needing to do `strlen()` on all the C-strings for atoms such as `{` `[`, `null`, `true`, `false` etc, as well as avoiding constructing temporary objects in some cases.  I was expecting a larger improvement -- but ultimately 5% is good enough as a first pass.

Test Plan:
---

1. First, build master (without this commit), and run the following:

   $ `ninja src/bench/bitcoin-bench`

   $ `src/bench/bitcoin-bench -filter=Block.\* -evals=20`

   *Note the min, max, median times as well as the total time.*

2. Next, build using this commit, and do the above again:

   $ `ninja src/bench/bitcoin-bench`

   $ `src/bench/bitcoin-bench -filter=Block.\* -evals=20`

You should see about a 5% speedup in the code using this commit (2)
above.

Correctness is not affected -- I tested and JSON outputted from bitcoind looks the same, and I ran a synch against Fulcrum and it is all correct (no errors) -- although I welcome any reviewer to read the code as well and provide feedback.