[lint] Fix Python script linting transgressions
At time of BCHN creation, we disabled the Phabricator-based linting.
We are selectively reactivating it with the introduction of a static_checks
CI stage and re-introduction of the tools and configuration.

This is the first in a series of language-specific commits which repair
some small transgressions w.r.t. the existing (still ABC) rules which we
will observe until further notice.