Strip trailing / in path for git-subtree-check
git-subtree-check fails if the directory is given with a trailing slash,
eg:

```
> test/lint/git-subtree-check.sh src/univalue/
ERROR: src/univalue/ is not a subtree
```

Shell autocompletes will add the trailing slash when autofilling the
path name, which will therefore cause the script to fail.

Just ignore any trailing slash.

(backport of Core commit 60582d6060542c1e3a23141ea825e36818fbbd54 as
applied to git-subtree-check.sh only, as part of reverting the script
removal and bringing it up to sync with Bitcoin Core version)

Signed-off-by: freetrader <freetrader@tuta.io>