Update the PPA to support Ubuntu 20.04, drop 16.04
Summary:
The new Ubuntu LTS version 20.04 "Focal Fossa" is about to be released
(scheduled on 2020/04/23). This diff removes the old 16.04 version and
add the new 20.04 version to the `debian-packages.sh` script, so the PPA
will serve the next release for Focal.

Test Plan:
  ../contrib/release/debian-packages.sh --dry-run <your_email>
Check there is a focal source.changes file generated.

Reviewers: #bitcoin_abc, deadalnix, jasonbcox

Reviewed By: #bitcoin_abc, deadalnix, jasonbcox

Subscribers: jasonbcox, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D5802

Commiter's note: BCHN is still providing supports for Ubuntu
Xenial 16.04, in fact the end of standard support by Canonical
will be on April 2021.