Add CLI option -?? as shorthand for -help-debug and make it imply -?/-help/-h
This MR makes it possible to simply use `-??` or `-help-debug` instead
of the unnecesary long `-? -help-debug` or `-help -help-debug`. The new
`-??` option is also listed directly below `-?` at the top of the CLI
options list, where it is much easier to find. The long form
`-help-debug` becomes a hidden option, just like `-help` and `-h`
already are.

This MR supersedes the need to backport
[PR14207](https://github.com/bitcoin/bitcoin/pull/14207)/[D5486](https://reviews.bitcoinabc.org/D5486).
The new behaviour of `-help-debug` is equal to current Bitcoin Core and
Bitcoin ABC; our availability of `-??` is the only behavioural
difference in the new situation. Apart from that, only the help text
differs.

Test plan:

1. `bitcoind -?` displays simple help
2. `bitcoind -??` displays help including advanced options
3. `bitcoind -help` displays simple help
4. `bitcoind -help-debug` displays help including advanced options
5. `bitcoind -help -help-debug` displays help including advanced options
6. `bitcoind -h` displays simple help