RPC performance enhancement for getmininginfo
Summary
---

This is in the spirit of keeping RPC code as efficient as possible. I
noticed when backporting D5056 and D5373 that the `getmininginfo` RPC
call was not transitioned over to a faster scheme.  As such, I went
ahead and created this MR, which depends on !251 and adds to it.

Test Plan
---

```bash
ninja check
ninja check-functional
```