Improve the listtransactions test for labels
This extends the test case for the feature to list transactions by labels
(using `listtransactions`) which was merged in !249 under the condition
that the test would be improved.

wallet_listtransactions now tests that:
* an unknown label returns empty result
* wildcard label works correctly
* tests for specific labels return the specified results

The change to assert_array_result() was needed to provide the functionality
to correctly match. Without the change a KeyError would be raised when
trying to filter for a key not present - however the method should
treat items which are missing a specified search key as non-matches.

Test plan:
* `ninja check-functional-extended` (because of the util.py change)