Minor performance fix for uint256 (base_blob) type
This also affects the uint160 type.

Summary of changes:

- Used a constexpr noexcept default constructor (this makes things like
  std::vector of this type potentially faster).
- Got rid of redundant declaration of constructors in subclasses,
  in favor of modern C++ using to inherit constructors
- made base_blob::size be static constexpr, this makes static_assert using
  this type possible.

The new changes allowed the compiler to deduce a new unused instance of
this class -- which was in src/script/sign.cpp.  This unused variable has
been removed in this commit.

In modern C++ you can do foo.method() as well as Foo::method() if
method is static constexpr, so this change will not negatively impact
extant code.