[DEPENDS] Add jemalloc to the depends
Summary:
This allow to build jemalloc as a static dependency. It also fixes the
missing dependencies for a successful static build on Posix platforms.
Next step is to enable it by default and use it for our release build.

Test Plan:
  cd depends
  make build-all

Cross build for all platforms with jemalloc enabled and run the
binaries (requires D6055).

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Subscribers: deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D6038