[CMAKE] Move the OpenSSL symbol detection to Qt rather than config
Summary:
This is a follow-up for the comment
https://reviews.bitcoinabc.org/D6243#inline-39530

The symbol is tightly coupled to OpenSSL, which only use it for BIP70
support in the GUI. This diff moves the check in the GUI where it
belongs and sets a preprocessor definition instead of setting the
definition in the config header.

Test Plan:
  ninja all check-bitcoin-qt
Cross build for all platforms.

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Subscribers: deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D6251