[backport] bitcoin-cli: Eliminate "Error parsing JSON" errors
Just pass arguments that can't be parsed as JSON as string arguments. If the
RPC method isn't expecting a string argument, this will result in more specific
errors from the RPC. And if it allows an argument to be a string as one
possible type (for example the `hash_or_height` argument to `getblockstats`),
this avoids the need for bitcoin-cli users to use double JSON quoting+shell
quoting to pass the string.

This is a port of https://github.com/bitcoin/bitcoin/pull/15448

Includes several adaptations by Andrea Suisani:

- Remove unused ParseNonRFCJSONValue and related unit tests
- Avoid double inversion in the same if condition
- Functional test adaptation

Merge #15488: bitcoin-cli: Eliminate "Error parsing JSON" errors

Test plan:

- On a version without this patch (!446), run `bitcoin-cli getblockstats 000000000019d6689c085ae165831e934ff763ae46a2a6c172b3f1b60a8ce26f` (passing the hash value without quotes ), and observe you get the following error message from the command line program:
  `error: Error parsing JSON:000000000019d6689c085ae165831e934ff763ae46a2a6c172b3f1b60a8ce26f`
- On a version with this patch, run the above command and you should get block statistics JSON output instead of the error message.
- Run `test/functional/test_runner.py interface_bitcoin_cli_getblockstats` - it should pass.