BanManager: Reduce costly linear scans & modernize code
As was observed in MR !497, the ban manager is inefficient in that it
does O(N) lookups for each IsBanned() query.  It must do this because
the ban table was being stored as a generic mapping of CSubNet -> CBanEntry.
CSubNet is a "catch-all" class that can encapsulate either a subnet or a
an IP address (an IP address is a special case of a subnet with all of
the netmask set to 0xf).

The reaason why previous to this MR it did a linear scan is because
for each IP address in question, it needed to decide if the CSubNet matches
the IP by calling CSubNet::Match.  But most of the time, though, the CSubNet
was just basically an IP (ip4/32 or ip6/128).

So, we decided to separate out the (potentially unbounded) per-IP
address bans from the subnet-based bans into two tables.

Note that in this codebase most of the time the ban table is created by automatically
banning individual misbehaving nodes by IP and thus we can benefit from storing
these two tables:

 - a per-IP address table (which is expected to grow potentially large), but
   which we will consult using O(1) constant time operation.
 - and as a fall-back, a per-subnet table (which is only ever manually edited
   by the user explicitly).

The sub-net table can be scanned linearly as before, but the per-IP address
can benefit from operator== and std::hash to store bans in a
std::unordered_map for constant-time O(1) lookup for IsBanned().

This commit closes issue #122 and allows us to merge !497 without reservation.

Notes:

- In order to maintain backward compatibility, we serialize the two ban tables as
  before, "as-if" it were a single unified std::map. This is acceptable
  performance-wise and will cause no file format breakage, but it did
  require us to implement serialization methods for the new BanTables
  class (this class replaces banmap_t)
- On deserialization we separate out the IP addresses from the SubNets
- This commit appears large because we also modernized some of the C++
  code in netaddress.cpp and also got rid of lots of double-copying and
  other inefficiency.