Expand MurmurHash3 API to accept any data type, not just std::vector
Summary
---

The MurmurHash3 function is a very fast hash function suitable for use
in e.g. things like std::unordered_map as a hasher. It is not
cryptographically secure, which is fine for a generic hasher for use in
internal hash tables (since it does support seeding).

I found myself wanting to use this hash function in other parts of the
codebase but I couldn't because the API presented is very inefficient.
It currently only accepts a std::vector<uint8_t>, which is oddly
specific for a generic hasher.  Not all data you want to hash is already
in a vector (such as data in CNetAddr, for instance, or perhaps a simple
int array, etc).

Requiring the caller pass a vector leads to redundant, useless copies on
the stack to create a temporary vector object -- which is potentially
wasteful of CPU cycles if the object you are hashing is large.

So, this commit adds an additional overload for this function, passing
it a generic const uint8_t * buffer pointer with a size.

This code introduces no semantic or externally observable behavioral changes.
The core of the hash function is the same code -- adapted to accept generic
uint8_t * pointers.

Notes
---

- This MR is optional.  If you feel it may create backport burden feel
free to close it.  However -- I would argue being able to use
MurmurHash3 in any code -- not just code that uses std::vector<uint8_t>,
is useful.

- Most of the low-level data hashing functions and other data
mogrification functions in this codebase accept both C-style data blobs
as well as std::vectors.  MurmurHash3 was an odd exception in that regard.

Test Plan
---

`ninja check`

The C++ test suite already tests MurmurHash3 thoroughtly using the existing
std::vector API -- which now internally calls into the new low-level C-style
API anyway.. so ninja check is sufficient to ensure it all works.