Performance: Optimized HexStr in util/strencodings.h
Optimized the heavily-used function `HexStr` which is used to encode binary data to hex strings e.g. `"ab01ef001234"`, etc.

This is part of our on-going effort to address #51 and make JSON-RPC much faster.

Motivation: This function ends up being called a lot and used a lot, expecially in the RPC server JSON execution paths. I ran a profiler on Linux and it revealed a good bit of time was spent in this function when doing
e.g. `BlockToJsonVerbose` or other RPC that encodes hex.

The following steps were taken to make this function much faster:

- Factored out the two `fSpaces` vs `!fSpaces` use-cases outside of the inner
  loop to reduce branching while looping.
- Don't use string.push_back, since it does excessive checks and
  branching.  Instead, pre-allocate the entire string and write to its
  data directly. We know the exact size of the resulting hex-encoded
  string ahread of time, so we can take advantage of that fact and avoid
  using push_back.
- We made the lookup table be for the full byte (and moved it to a single translation unit). This results in fewer bit-ops and bit-shifts and saves a few cycles in my testing.  We copy the data as 2-char pairs at a time, rather than tearing the byte up into hi/low nibble.
    - Note that the hex lookup table grew from 16 bytes to 512 bytes -- HOWEVER -- now it lives in only 1 place in the program (previously it could potentially live in more than 1 place since this was a template function with a static table).. so it's possible the memory and binary size footprint is the same as before.

Running a benchmark on `BlockToJsonVerbose.\*` reveals a 10-20% speedup
for the 1MB and 32MB block benchmarks.

Comment: This function has become slightly harder to read. However, given how critical it is for performance, it is my opinion that this small cost is ok given the performance benefits conferred by these optimizations.

**Update:** Reading the code again I realized the original function in master has a bug whereby invalid usage
of the function (begin > end) could lead to an app crash via `std::bad_alloc`.  It wasn't checking its inputs.
It could be argued this is undesirable.  I added a test case for this situation and now the function enforces
that begin <= end.  If that's not the case, it always returns an empty string now (as opposed to crashing the app). I pushed a commit to address this.

Test Plan
---------

```bash
$ ninja check check-functional-extended
```

Note that there are extensive and very thorough tests in `ninja check` for this exact function to ensure its correctness (see: `tests/util_tests.cpp`).

If you want to see the performance gains:

1. Switch to `master`.  Do:

```bash
$ mkdir -p ~/tmp/
$ ninja bench_bitcoin
$ cp src/bench/bench_bitcoin ~/tmp/bench_before
```

2. Switch to this branch and do:

```bash
$ ninja bench_bitcoin
$ cp src/bench/bench_bitcoin ~/tmp/bench_after
```

Then run the following and compare results:

```bash
$ ~/tmp/bench_before -filter=BlockToJson.\*  # this is the master build, should be slower
$ ~/tmp/bench_after -filter=BlockToJson.\*  # this is this MR's build, should be 10-20% faster
```