Always repport proper sigcheck count
Summary:
This remove some activation trick that was used for sig checks.

Now the interpreter will always return the proper sig check count, but that count is ignored when connecting older blocks which may have signature count in excess of the limit.

Test Plan:
  ninja all check-all

Run IBD on testnet and mainnet

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D6567