Follow-up to MR !529 -- added better comments/naming
The workaround to the WSL bug that is MR !529 only affects WSL v1
systems.  I verified this myself on a WSL v2 system.  As such, it is
better that the code be clear and explicit about what is affected and
what isn't.  So I renamed the function `IsWSL` -> `IsWSLv1` and I
modified the comments to be explicitly clear that the workaround is only
for v1.