Improve CRollingBloomFilter API to avoid double copies
Summary
---

This MR depends on !541 and is a commit on top of it. The
CRollingBloomFilter class takes a std::vector<uint8_t> as its input and
this is slow if trying to use the bloom filter with arbitrary data since
it requires the caller to create a vector.

In future work, we want to start using this class with BanMan, with
the CNetAddr class, which does not store its address data as vectors. So using
the CRollingBloomfilter would involve lots of double copies to just insert
and/or test set membership, which is not the end of the world, but
is not ideal either.  Given how !497 may eventually need to be merged
(to test bloom filter set membership for every vAddr being relayed to
peers -- see !497) -- we do not want to do double copies each time we
test set membership.

What's more, the API accepted a uint26 hash but then it would internally
copy it to a vector.  This is just unnecessary.

So, instead, the API has been changed: it now accepts const uint8_t *
and a size as the fundamental canonical way to insert/test the bloom
set.  It leverages the MurmurHash3 improvements from !541 to accomplish
this.

The existing std::vector and uint256 APIs still exist but they just call
into the common uint8_t */size functions to do the actual work.

The bloom tests were modified to also test this new API.

Test Plan
---

`ninja check`

The existing bloom_tests.cpp should be sufficient, since no semantic or
behavioral changes are introduced by this commit aside from the API being
faster for general use.