UniValue: Fixed 2 small bugs, plus small performance improvement
Summary
---

The univalue_read.cpp getJsonToken() function had two small bugs:

1. If the buffer being parsed is just a single hyphen (that is,
the C-string char '-' and then '\0', it would incorrectly accept that
as a VNUM, rather than emitting a parse error. This would leave the
instance in an inconsistent state (VNUM with data "-" is invalid JSON).

2. If the buffer being parsed was the bare number '00' or '-00', it
would read-past-the-end of the buffer (into the terminating NUL byte).
In practice this didn't matter because there was always a terminating
NUL byte to read, because we are ultimately reading from a std::string
which guarantees a NUL byte at the end -- however if getJsonToken()
should ever be used with anything other than a std::string as the source
buffer, it might fail with a seg fault or undefined behavior. This has
been addressed.

Unit tests were added to handle these 2 situations, plus some other
extra tests were added as well.

In addition, this commit reduces copies & mallocs by using std::move()
in 1 place, and in another, by not pushing each number's digit at a time
but rather copying the whole number string once the buffer is accepted.

This commit makes the JsonReadWriteBlock_32MB bench run 10% faster on my
system.

Test Plan
---

- ninja check-univalue
- ninja check-all

There are no semantic changes other than fixing the issue where bare
'-' was incorrectly being accepted before (now it's being rejected, as
it should be).