[ci] Enforce RPC coverage in CI
Follow-up of !571

This coverage checks the RPC functions. The CI will complain if any RPC
methods are not tested.

This creates a job that runs the full extended suite of tests, instead
of the setup we have now here quick tests and long tests run in two
separate jobs. This is required to be able to test full coverage.