timedata.cpp nits: avoid extra copies, use thread safety macro
Summary
---

This is another nit commit that does not change the code's behavior in
any way but simply makes the code leaner and meaner and also more
defensive.

- the .sorted() method of CMedianFilter returned a *copy* of the
internal vector. This caused extra copying to happen when none was
needed. In the way this class is being used (it's only used in
timedata.cpp), it's better to return a reference.  The vector won't
mutate anyway so this is safe.

- Added thread safety decorators for the stuff in timedata.cpp that
should be guarded with the lock to catch any future modification
of this code breaking thread safety. Tested on clang, works.

- Updated a test to check the .sorted() method's correctness.

Test Plan
---

- review the code
- ninja all check