wallet rpc: Fix to unused parameter caused by not passing coinControl down
Summary
---

While reviewing !617, I noticed the commit is a step backwards in terms
of code quality.  The static `SendMoney` function is called in only one
place and they pass a `CCCoinControl` instance to it which is default
constructed.  Fair enough.  However.. inside the function they never
used the passed-in value (again, this is called in only 1 place), but
instead create their own CCoinControl *again* to pass down to
`pwallet->CreateTransaction`. Aside from generating warnings about
unused paramters -- this is just clearly an error on the original
committer's part.

This commit addresses that concern.  The issue was raised here:

https://gitlab.com/bitcoin-cash-node/bitcoin-cash-node/-/merge_requests/617#note_377169795

In addition the original commit was using snake_case mixed with
camelCase which looks odd considering everything else is camelCase. This
commit sticks to camelCase as that's what everything else is.

Test Plan
---

`ninja all check-all`

This commit is just a code quality/nit commit and introduces no changes
or semantic differences.