Remove UniValue::clear()
This removes the ambiguous UniValue clear() method, in favour of the
setNull() method which does exactly the same but has a ... clearer name.

For example, the fact that clear() changes an array into null instead of
an empty array may be unexpected, but with setNull() this is obviously
expected.

This fits within the code quality objective of #51.

Test plan: `ninja check-univalue check`