UniValue fix: Serialize floats properly, irrespective of locale
Summary
---

It turns out Qt was setting the C global locale for us (so nice of it to
do so), so Bitcoin-Qt ended up serializing JSON floats, if using locales
such as de_DE, using ',' as the decimal character.

I checked the C locale documentation for printf/sprintf and the decimal
character for floats is the only possible thing that we care about that
the LC_NUMERIC or LC_ALL locale settings can affect.

Since we can't manipulate the global C locale .. since we are a multi-
threaded application and the C locale is fundamentally static/globale, and
since it may even *differ* from the C++ std::locale, we have no choice
but to completely avoid the std::snprintf() function (which is faster
than std::ostringstream), and fall back to using std::ostringstream for
serializing doubles.

Benchmarks don't show much of a performance hit -- likely because
doubles in RPC are very rare in this code.

I also added setting the lcoale to "" to the UniValue tests as the first
thing the tests do before they run, so as to support CI integration
where the tests are run with LC_ALL or LANG env vars set to an alternate
locale.

This fixes bug #137.

Test Plan
---

First reproduce the bug:

1. Compile from master (not this branch)
2. Run bitcoin-qt with ENV var set: `LANG=de_DE.UTF-8`
3. Hit the RPC either from the Debug Console or via CLI and issue the
`getmininginfo` RPC command -- you should get an error if using the CLI
(the JSON being produced under the hood has commas instead of decimal points
in it!!), but if you are using the console you will see the JSON text
that is broken.
4. Run the above again without the `LANG=` var set and see it work.

Now, do 1-4 above again this time against *this MR*.  You should see
JSON output is unaffected and works regardless of LANG= env var.

After doing the above, do:

- ninja all check-all check-univalue

Just to be sure.