Fix unused function warning for `GetPubKey` in `script/sign.cpp`
Summary
---

The function `GetPubKey` in `script/sign.cpp` is unused currently
and is producing compile-time warnings.

We anticipate needing it again in the future. So for now, we are
therefore commenting it out to eliminate the warnings, but leaving
it in-place in anticipation of its future resurrection.

Test Plan
---

- `ninja all check`

Also when you do build from scratch, confirm that you do not see this
warning (this is on clang, gcc may produce something similar):

```
Building CXX object src/CMakeFiles/script.dir/script/sign.cpp.o
../src/script/sign.cpp:53:13: warning: unused function 'GetPubKey' [-Wunused-function]
static bool GetPubKey(const SigningProvider &provider, SignatureData &sigdata,
            ^
1 warning generated
```