Updated RPC help for getblocktemplate for 'sigops'
Summary
---

Also edited the help for the 'fees' key which was missing the word
'this'. The update to the 'sigops' key renamed the concept to talk about
sigchecks, which is what we report now, rather than sigops.

Re-generated docs to reflect these changes.

Test Plan
---

1. Review changes to both getblocktemplate.md and sources.
2. (optional) build bitcoind, use RPC: `help getblocktemplate`, review
help text.  In particular the per-text `sigops` field description has
changed, as has the `fee` description (added the word "this" to make the
English be grammatically correct).