Restrict CBlockIndex to not allow copy, move, or assignment
Summary
---

This commit restricts the linked-list type `CBlockIndex` to not allow
copy, move, or assignment by value.  This is done for type safety.

Motivation
----

It is a programming error to move, copy, or assign this type by value since
it is a linked list node.  Every instance of this type normally is created
on the heap and has a fixed location in memory, and other nodes point to
it using raw pointers.  Thus, copying or moving this type is dangerous
and should be disallowed on the grounds that if not done carefully, it could
lead to a dangling pointer situation.

Note for tests: Restricting this type to not allow move, copy, or assignment
also makes it impossible to create unit test chains of `std::vector<CBlockIndex>`
which is actually also a safety feature, since vectors may resize themselves,
thus leading to a situation with dangling pointers in unit test code.

As such, the existing tests that did this (and thus were very prone to
programming errors if modified) have been changed to not create these types on
the stack, but rather create them on the heap using `unique_ptr`s, thereby being
able to safely put instances of the pointer into a vector.  Note that the
core codebase of the application (almost) always puts CBlockIndex
instances on the heap, as this is the only real meaningful and safe
usase of this class (because it is a linked list node).

Also note that while developing the unit tests for ASERT, other developers on
this project were bitten by the danger with using this type incorrectly and with
copying it around or moving it around by value. Thus, such usages are clearly
error-prone and should be forbidden.

Test Plan
---

`ninja all check check-functional`

No behavioral changes were introduces into the code.  This change
merely restricts the `CBlockIndex` type to have a new set of more restrictive
rules for usage in the codebase -- thus if the code compiles and runs, we know
that no errors or bugs have been introduced with this change.