remove SCRIPT_VERIFY_CHECKDATASIG_SIGOPS flag
As remarked in its definition it is now a meaningless flag
now that sigchecks upgrade is complete (and sigops no longer
exist).

(Historically this flag was used to enable the opcode, then
after the CDS upgrade it became a control on the sigops counting
only. Now after sigchecks it is meaningless.)

Test plan
=========

`ninja check`