Made the CTransaction class no longer default constructible
Previously, the default construction of CTransaction was used sparsely
in this codebase -- mainly to capture the concept of a "null" or empty
transaction.

The code comments themselves warned that the default constructible
CTransaction class is not desired and indicated in a TODO that the
default construction should be removed.

I am in agreement with this -- for type safety and to prevent people
working with CTransaction vectors directly (really everything should be
using CTransactionRef).

So I made the default constructor private (and I also made the
CTransaction class final).

Instead we expose a global static member `null` and `sharedNull` to
capture the concept of "nullness" where it is needed (mainly in test
code).

The changes to the code were surprisingly minimal to accmplish this.

This commit introduces no semantic or behavioral changes to the code,
this is purely a code quality commit.