[txmempool] Remove dangling/obsolete code in removeForBlock(...)
In https://reviews.bitcoinabc.org/D2172, deadalnix removed the block policy
estimator code. This included one line in TXMemPool::removeForBlock which
used the `entries` vector (which contained the intersection of the block's
transactions and the mempool's transactions). Nothing else uses `entries`.
The code that creates the `entries` vector was left in place. It's not
needed for anything, so it should be deleted.

The for loop that generates `entries` consumes about 1.6% of the total
removeForBlock(...) runtime, so it's not a big deal in terms of performance.

Test plan:
ninja check-all

https://gitlab.com/bitcoin-cash-node/bitcoin-cash-node/-/commit/51f94e21e90308a3509404ed3b4e8fd063c11892#685ae4924603abb0150ce8b264d455a762658039_676_668