Implement Extversion
Summary
=======

This implements the `extversion` extended versioning handshake
protocol in a simple, minimalistic way. MR !753 is a continuation
of Greg Griffith's work from !558, with a lot of extra code,
C macros and some duplication removed.

This commit bases on the specification v0.1.0 authored by
Awemany and Griffith which can be found in `doc/xversionmessage.md`
or at https://docs.bitcoincashnode.org/doc/xversionmessage/ .

The extversion handshaking is disabled by default and can be enabled
using the `-useextversion` CLI argument or the `useextversion=1`
configuration file setting.

Nodes indicate extversion support by setting service bit 11 (EXTVERSION)

Differences from BU implementation
==================================

We do not use a map to store key/value pairs received from the peer
unconditionally, since storing random values we do not understand as
binary blobs does not have much value and incurs a cost (memory).
We avoid the possibility for peers to send us data we don't understand,
up to 100kB or so, which we would blindly store for the duration of the
connection. We instead use an approach where we only read the ExtVersion
keys we do understand and we place them into C++ types such as structs
so that they may be accessed from code directly.
This approach enforces type safety and also reduces the impact of Extversion
on the codebase.
In the future if we want to support very many keys, we should revisit this
design, since in some cases a std::map is more efficient (when empty) than
a bunch of unused structs, per peer.

Right now ExtVersion supports only 1 key -- the "Version" key.

Test Plan
=========

Basic:

- `ninja all check check-functional`

Advanced:

1. Start a testnet node (if you have one synched) with the CLI args -useextversion -debug=net.
   If you do not have one synched, you can skip this section.
2. Make sure to connect it to a BUCash Node that has "extversion" (not xversion).
   One such note is: tbch.loping.net:18333.
   You can do `addnode tbch.loping.net:18333 onetry` via RPC to add it.
3. Check the log -- you should see something like below if the extversion handshake succeeded:
   (in particular look for a message of the form `extversion: peer=xxx is using extversion 0.1.0`)

    ```
    2020-09-03T08:29:42Z receive version message: [183.88.131.118:18333] /BCH Unlimited:1.9.0(EB32; AD12; 64bit)/: version 80003, blocks=1407069, us=78.97.206.149:53352, peer=1
    2020-09-03T08:29:42Z added time data, samples 3, offset -2 (+0 minutes)
    2020-09-03T08:29:42Z received: extversion (172 bytes) peer=1
    2020-09-03T08:29:42Z extversion: peer=1 is using extversion 0.1.0
    ```

4. Alternatively, you can issue the RPC getinfo command and that extversion-capable peer
   I mentioned above should show you bytes sent/received for the network message extversion
   in the bytesrecv_per_msg and bytessent_per_msg sub-dictionaries.

5. You may also use the GUI and connect to an extversion-capable peer (e.g. BCH Unlimited 1.9.0+).
   You should see the GUI display EXTVERSION in its service bits
   (the window is formatted strangely and you may have to expand the window horizontally
   to see all the service bits).

Co-Authored-By: Griffith <probablyaplebeian@protonmail.com>
Co-Authored-By: Calin Culianu <calin.culianu@gmail.com>

Reference to BU PR: https://github.com/BitcoinUnlimited/BitcoinUnlimited/pull/1236