Fixed UniValue bug: don't take pointers to temporaries
This was a bug introduced by the merge of !772. The UniValue changes
introduced 2 new types, UniValue::Object & UniValue::Array both of which
can implicitly be converted to UniValue (as a temporary, in some cases).

Some of the code in the codebase was taking a pointer to this temporary,
and then on scope end the temporary would become invalidated, while the
pointer would remain.

This in particular was in the exeuction path of getblocktemplate
'longpoll' in rpc/mining.cpp.  We had CI failures related to this, thankfully,
off master, otherwise this may have gone unnoticed.

This commit addresses the situation -- I scanned for places where
get_array() and get_obj() were being called and I ensured that where
appropriate, we don't create any temporary UniValue instances from the
returned contianer type.

The crucial fix was in rpc/mining.cpp.  The other fixes are merely code
quality nits to avoid creating extra temporaries.