Prohibit implicit UniValue::VType to UniValue conversion
I'd say it's likely we'll eventually drop the `VType` constructor
completely (in favour of the type-specific overloads), but for now this
makes it explicit to discourage unnecessary use (such as `= VARR`
instead of `.setArray()`).

It turns out there is no such implicit usage at the moment, so a
compile-only test plan.

Test plan: `ninja all bench_bitcoin test_bitcoin check-univalue`