Remove declared but undefined function: fsbridge::freopen
This function is declared in fs.h under `namespace fsbridge` but there is no
implementation for it. Obviously, nothing calls it either.

In the interests of code quality it should be removed, lest someone
write code expecting it to exist, only to get angry later that it does
not.  If we find we ever need it we can go ahead and add it again and
this time actually implement it.