Replaced many instances of std::string pass-by-value with pass-by-ref
Summary
----

This is a code quality and performance nit MR.

In many places in the code, arguments to functions were taking
`std::string` by value rather than by `const` reference.  Not only is this
not idiomatic, but it can eat CPU cycles as extra copies are created in
some cases.

I ran through the code and did a regex search for such usages, and I
then audited and reviewed each place that turned up a match.  All of the
changes here are safe and are preferred, to avoid the potential for
excessive copies.

In a couple of places I also caught a few other args that were not being
passed by reference but rather by value-copy.

There tended to be a pattern where if `std::string` was being passed
incorrectly by value, so were the other args to the function in question.

In a few places in the code pass-by-value was being done with intent;
sometimes the code would modify its argument in-place for its computation,
or the code in question would immediately `std::move()` the string to a
member variable.  In such places the code was obviously not modified.

But the majority of hits were in fact incorrect usages that risked
the app doing extra copying of the string where single copies or no
copies would be done if `std::string` were passed by `const` reference.

In at least 1 place I also changed incorrect `push_back(ValueType{arg1, arg2})`
to `emplace_back(arg1, arg2)` which avoids the potential for extra
temporaries created.

Test Plan
---

No semantic changes were introduced but I encourage reviewers to read
the code and double-check.  I went back and reviewed each change myself
carefully.  In general there is no possibility that this creates
problems because the type system and compiler catches all gotchas here
(this is a case where the type system 100% works in our favor).

To be sure, compile and test:

- `ninja all check-all`
- `ninja bench_bitcoin && src/bench/bench_bitcoin -evals=1`