Make chainparamsseeds.h use numeric strings for human-readability
This commit adds support to the `SeedSpec6` class to support parsing
numeric IPv4 and IPv6 host:port notation directly in the code.

It was found that chainparamsseeds.h is completely unreadable by humans.
It is useful to be able to read the seeds strings when looking over the
codebase.

While one could resort to opening up the text files in contrib/seeds/nodes_*.txt,
it's not quite as convenient as being able to read the header and scan over
the seed ip:port pairs visually.

As such, this commit adds that capability.

- Unit tests were added for this new facility that are based off the existing
seed list (using array octet notation) before this commit -- comparing
the parsed host:port pair with the hard-coded expected bytes.

- The `generate-seeds.py` script was modified to produce output in the
new format.  It still validates its inputs from nodes_*.txt, but now it
produces arrays that are human-readable strings, rather than arrays of
bytes.