Scalenet
Testnet3 has gotten rather bloated and takes an annoyingly long time to sync.
This happened because we did not have an appropriate place to test node performance
in high-throughput, spammy situations.

This MR creates scalenet, which is intended to be the designated place for spam once
we reach block height 10,000.
Every 6 months or so, scalenet's block 10,000 will be invalidated and a new block
will be checkpointed in its place, clearing out all of the spam history, and keeping
scalenet semi-affordable to sync.

Scalenet is intended to target the performance level of a ~$40/month VPS or a $500
desktop computer for the near future. Any tests that target higher performance levels
are encouraged to do so by forking off of scalenet or creating their own private testnets
or regtest networks.

See also:

1. https://bitcoincashresearch.org/t/testnet4-and-scalenet/148
2. https://gitlab.com/bitcoin-cash-node/bitcoin-cash-node/-/merge_requests/739

# Test plan

 * `ninja check`
 * `bitcoind -scalenet` and make sure it fully syncs

Co-authored by: Calin Culianu <calin.culianu@gmail.com>