[rpc] Replace error heuristics with rc check
Replace heuristics in SendMoney for diagnosing why creating a
transaction failed with checking return value of CreateTransaction.

Jonathan Toomim discovered in !442 that replacing the balance check
in `sendtoaddress` resulted in a 13% speed improvement.

In review, it was discovered that if the return value of
`CreateTransaction` was improved, the check could be removed altogether.

Removing the check does not improve the speed significantly as compared
to the improvement implemented by Jonathan !442, but simplifies the
code.