RPC: Alias `dummy` for second arg to `submitblock` for compatibility
Summary
---

As a result of the `getblocktemplatelight` MR !281, we ended up renaming
the (ignored) second argument to `submitblock` from `dummy` to
`parameters`.  Since it was recently discovered that renaming arguments
can sometimes cause problems for software in the wild (see: issue #173),
we may wish to err on the side of caution and keep RPC API compatibility
by keeping older parameter names aliased to the new names we have given
them.

This commit does just that -- it makes the second arg to `submitblock`
accept the named parameter `dummy` or `parameters`. This maintains 100%
compatibility with other APIs such as Core or ABC.

The `mining_basic.py` test was updated to test that the alias works and
is supported.

This MR is similar in spirit to !810.

Test Plan
---

- `ninja check-functional`