Fix EXPECTED_VIOLATION_COUNT warning for `check-functional`
Summary
---

A recent commit must have not set this value properly.  Here is the
warning message when you do `ninja check-functional`:

```
INFO: 27 tests not meeting naming conventions (expected 26):
  abc-cmdline.py
  abc-finalize-block.py
  abc-get-invalid-block.py
  ...
```

This commit fixes ths situation by bumping the number in
`test_runner.py` from 26 to 27.

Test Plan
---

1. `ninja check-functional` both with and without this commit.  You do
   not need to even let the tests complete. You can ctrl-c after a few
moments.  The only thing to verify is that without this commit you get
the warning and with you do not.

2. If you like, do let `ninja check-functional` complete with this
   commit just for fun.

No semantic or anything changes are introduced.  This is a 1-character
change.