Fix `ninja check` failure on Arch Linux
Summary
---

This fixes issue #175. The underlying problem is that locale C.UTF-8
does not exist on some systems, notably Darwin and Arch Linux.  On Darwin
C.UTF-8 is silently accepted but on Arch it is not, and it produces an
error/warning when running the tests `ninja check`.

On Ubuntu-based systems and Fedora "C.UTF-8" is a real locale, so this
commit preserves existing behavior on those systems.  Only on Arch this
commit should fix ths issue.

Test Plan
---

`ninja check` should work on Arch now without any issues.  If you lack an
Arch system, just run `ninja check` as usual to confirm that no regressions
were introduced.