Fix ODR violation leading to subtle build error on Fedora
Summary
---

See the description in issue #176. This closes #176.  The basic jist of
it is that we inherited some questionable code from Core in which
mapOrphanTransactions was declared inconsistently and this caused newer
GCC to emit bad code and/or complain.

The fix is to actually declare things only once in 1 place and not have
such footguns waiting to go off in the codebase.  See diff.

Test Plan
---

- `ninja check check-functional`