[trivial] Remove extraneous debug print in mininode.py on_version
Summary
---

It was pointed out by one of the other maintainers that mininode.py has
an extra debug print statement which was introduced in this commit:

https://gitlab.com/bitcoin-cash-node/bitcoin-cash-node/-/commit/46dd8b2870dfa6c17b8a916d2b5258876e2b4758#213b7b9875934a9039752454f0ee8d030b2fb75a_403_405

This was likely a holdover from the original programmer doing some
debugging when adding the extversion facility.

This commit removes this line of code.

Test Plan
---

`ninja check-functional`