Make -maxmempool default to 10x excessiveblocksize (e.g. 320 MB on mainet, 2560 MB on scalenet)
# Overview

Currently, the default setting for -maxmempool is 300 MB.
This allows 300 MB of RAM consumption by the mempool.
Since each tx uses about 4.5x more RAM than its serialized size,
a 300 MB mempool allows for only about 64 MB of transactions in it.
This is borderline sufficient at 32 MB per block, but is insufficient
for larger block sizes (e.g. scalenet's 256 MB). This MR changes the
default value for -maxmempool to be 10 times the `-excessiveblocksize`
setting, whether that is the default setting (32 MB for mainnet, 256 MB
for scalenet) or set manually as a command-line option. This MR does not
change the semantics of the -maxmempool setting when it is set manually --
`-maxmempool=300` will still result in a 300 MB budget for the mempool,
same as before.

Due to the use of an integer multiple, this MR results in a slighly
different default `-maxmempool` setting on mainnet:
320 MB instead of 300 MB.

# Test plan

ninja check-all