RPC: Make `getblockstats` not require txindex, use undo file instead
Co-authored-by: Calin Culianu <calin.culianu@gmail.com>

This is a (partial) backport of Core: #14802. The reason why it's not a
cherry-pick or a more full backport is that their commit didn't cleanly
apply on top of our pre-existing optimizations to `getblockstats` and
had to be significantly adapted.  It also contained updated block .json
data that was segwit which we can't use, so we opted not to use it.

Their commit also broke compatibility in a corner case: they no longer
allow getblockstats for block 0, which the previous RPC did allow. We
preserve 100% API compatibility with the previous RPC by allowing
getblockstats on block 0.

Original commit message:

Using undo data for a block (rev?????.dat) we can retrieve value information
about prevouts and calculate the final transaction fee (rate). This approach
is about 80x faster, drops the requirement for -txindex, and works for all
non-pruned blocks.