Fix for compile problem in rpc/mining.cpp on some platforms
Summary
---

Addresses issue #195. Recent change in MR !867 boldly attempted to go
where no return type has gone before in this codebase and use `auto` and
some compilers were not happy.  Clang on Darwin namely. See issue.

Test Plan
---

- `ninja check check-functional`
- Optional: If you have access to a Darwin system, ensure it compiles
  there.  I am not sure if this problem also manifests itself on clang
on Linux, so maybe it would be nice to try building on clang on Linux
too as a check as well.