[qa] fix unreliable wallet_basic tests
Summary
---

This fixes issue #186.

Commit 3f936a1c2e70d699ea9609d7cbea0f02e99828cf introduced unreliability
into the wallet_basic tests because there is a low chance of node 0
ending up with the 0.000002 TX fee of the TX sent by node 1.

To fix this, we move the test and let node 1 generate a block to ensure
it gets its own TX fee back.

Test Plan:
---

* Run `./test/functional/test_runner.py --failfast -j $(nproc) $(echo {1..32} | sed 's/[^ ]*/wallet_basic.py/g')` and verify it doesn't fail.