Upgrade UniValue code in core_write.cpp (#51, #89)
Brings `core_write.cpp` in line with latest #51 tech.

Some refactoring occurred:

* `ScriptToUniv`, `ScriptPubKeyToUniv`, and `TxToUniv` now return a new
`UniValue::Object`, instead of modifying an existing container. For
convenience, the `in_active_chain` key is moved from the beginning to
the end of the object resulting from the `getrawtransaction` RPC (this
should not break any API client, since no client should rely on the
order of key-value pairs in dictionaries).
* `ScriptToUniv`, `ScriptPubKeyToUniv`, and `TxToUniv` are extended with
a `const Config&` argument. This is inserted at the front of the
argument list, so optional arguments at the end remain possible.
* Prepares fixing #89. In the future, `bitcoin-tx` can easily switch to
its own configuration object when removing `GetConfig()`.
* The JSON/RPC benchmarks now make use of `GetConfig()`, like other
existing benchmarks. The entire benchmark suite can be upgraded similar
to `bitcoin-tx` at a later time.
* An optional arugment `fIncludeP2SH` was added to `ScriptPubKeyToUniv`.
This is used by the `decodescript` RPC, so that it no longer needs to
reverse-engineer the result of `ScriptPubKeyToUniv`, which I declare
considered harmful.
* The `hex` field was removed from the RPC help text of `decodescript`,
because this is actually not part of the output (it's intentionally
omitted since it would be a copy of the input).
* `core_read.cpp` and `core_write.cpp` no longer include `univalue.h`,
because it is already included in `core_io.h` now (needed for
`UniValue::Object`).
* A SegWit reference was removed from a comment in `script/standard.h`.

Test plan: `ninja bench_bitcoin check check-functional;
src/bench/bench_bitcoin -evals=1`