Consistently name the "excessive block size" concept + add a test
Co-authored-by: Calin Culianu <calin.culianu@gmail.com>

Summary
---

Throughout the codebase, there are two concepts of max block size:

- "excessive block size" - (config arg `-excessiveblocksize`), which is
  the absolute largest block we accept from the network (consensus rule,
  32MB on mainnet). This size must simply be >= 1MB.
- "block max size" - (config arg `-blockmaxsize`) the block size we use
  for mining, if we are a mining node (used by the `BlockAssembler`, etc).
  This size must always be <= `-excessiveblocksize`.

The two concepts are sometimes muddled in the codebase in terms of naming
-- and in particular the Config object calls the "excessive block size"
the confusingly named `GetMaxBlockSize()`, which I fear might be a source of
confusion for maintainers of this codebase going forward, since it can
easily be confused with the "block max size".

This commit does the following:

- Renames the `Config` object getter/setter method for the "excessive
  block size" from `[Get|Set]MaxBlockSize()` to the much clearer
 `[Get|Set]ExcessiveBlockSize()`
- Renames the `DEFAULT_MAX_BLOCK_SIZE` constant to
  `DEFAULT_EXCESSIVE_BLOCK_SIZE`
- Renames the instance variable in `Consensus::Params` from `nDefaultMaxBlockSize`
  to `nDefaultExcessiveBlockSize`
- Renames the constant also in the test framework from
  `DEFAULT_MAX_BLOCKSIZE` to `DEFAULT_EXCESSIVE_BLOCK_SIZE` for
  consistency.
- Adds some appropriate comments to `config.h` and in a bunch of places
- Adds a test that was submitted to me by Mark (hence why he is named
  the primary author of this commit) -- that does some paranoia checks
  on his sigchecks feature.

I apologize for this large commit. Do not be intimidated by it. It is a
strict rename and the C++ compiler can be trusted here to catch any
errors in the refactor due to its trivial nature.

Also the additional test will do no harm and perhaps only help things.

Test Plan
---

- `ninja check check-functional-extended`