Mininode: Speed up big message reception significantly
Summary
---
The current mininode code is quite slow when receiving big messages.
This is caused by appending to the immutable `bytes` type instead of
using a `bytearray`.

This also changes the `data_received` handler to exit early without
acquiring the `mininode_lock` if the buffer is empty.

Another optimization is to move extracting the `command` and `checksum`
of the message to after the whole message has been received.

With this changes the time to receive a 256 MB scalenet block is reduced
from minutes to seconds.

Test Plan
---
* `test/functional/test_runner.py -j$(ncpu)`