[qa] fix unreliable getblocktemplate(light) tests
Summary
---

This fixes issue #204.

The tests for getblocktemplate were unreliable because there was a
slight chance that block templates differ by one second in `curtime`.

Instead of using mock time, we fix this by introducing a new helper
function, `assert_blocktemplate_equal`, which checks that the templates
are equal but allows a difference of up to one second in `curtime`.

Test Plan:
---

* Run `./test/functional/test_runner.py --failfast -j $(nproc) $(echo {1..32} | sed 's/[^ ]*/bchn-rpc-gbtl-bg-cleaner.py bchn-rpc-getblocktemplatelight.py mining_getblocktemplate_errors.py/g')` and verify it doesn't fail.
* Note that `bchn-rpc-gbtl-bg-cleaner.py` still fails sometimes due to a different issue, see #205