[qa] Add RPC HTTP pipelining test
Summary
---

The RPC server supports HTTP 1.1 pipelining which is not covered by any
test.

To address this we add a test that directly sends multiple pipelined
HTTP requests using plain Python sockets. We do this with multiple
requests in a single TCP packet and with a separate packet for each
request.

This was successfully tested with libevent version 2.0.22, 2.1.6, 2.1.8
and 2.1.12. I tested version 2.1.6 with and without the libevent
workaround, but in no case RPC was ever hanging or crashing.

Test Plan
---

* Run `test/functional/test_runner.py --failfast -j $(nproc) $(echo {1..256} | sed 's/[^ ]*/bchn-rpc-pipelining.py/g')` and verify it doesn't fail.