Nit: Fix clang-11 warning: "uninitialized variable as const referenece"
Summary
---

I tried clang-11 to see if it gets any new nits/bugs in our code. We are
all green except for in 2 places it doesn't like our use of
uninitialized variables being passed as a const reference.

Other compilers may dislike this type of thing as well.  In the
interests of code quality I decided to go ahead and fix these warnings.

Test Plan
---

- `ninja check bench_bitcoin check-univalue check-secp256k1 check-functional`

Optional: If you want, obtain clang-11 and try it against master versus this MR.
The warnings spewed forth on master should be gone on this MR.