[qa] Fix for function `hash256` appearing twice in test framework
Summary
---

This commit was the result of an issue raised in Slack.

There are two instances of a function named `hash256` in the test
framework, and they each have different semantics.  The one in `util.py`
returns a *reversed* sha256d (ready for hex display), whereas the one in
`messages.py` returns a sha256d (non-reversed).

This is a potential foot-gun since the two functions have the same name
but different semantics, and as such it should be addressed.

The one in `util.py` has been renamed to `hash256_reversed` and call
sites were updated for the 1 file in which it was used, `interface_zmq.py`.

Test Plan
---

- Review
- `ninja check-functional-extended`