util: Added functions to help with profiling code (more precise time & more)
Summary
---

This commit adds facilities to make profiling code a bit easier and more
precise.

- Added a new function, `GetPerfTimeNanos` which returns a steady monotonic
timestamp in nanoseconds, with sub-microsecond precion (on most
platforms), suitable for profiling.
  - This timestamp is on a completely different clock than the system
    time and should only be used to profile code.
  - On major platforms we support it will provide a monotonic steady
    clock with sub-microsecond precision.
  - For platforms where that is not possible, it falls back to
    GetTimeMicros() (which is the current status quo, based on the
    system clock, relies on boost, is not "steady").
  - It is hoped that this function can increase the resolution and
    precision for profiling code.  In future MRs or future research on
    performance bottlenecks, perhaps it can be used.

- Also added a class called `Tic` which wraps this functiom.  Tic
captures a timestamp from when it was constructed (via a call to
`GetPerfTimeNanos`).  It can be used to reduce the amount of boilerplate
code one writes when printing profiling output to the log.  For
instance:

    auto t0 = GetTimeMicros();

    // do some stuff

    LogPrintf("Stuff took %1.3f msec\n", (GetTimeMicros() - t0) / 1000.0);

    auto t1 = GetTimeMicros();
    // do more stuff

    LogPrintf("More stuff took %1.3f msec, %1.3f msec total time elapsed\n",
              (GetTimeMicros() - t1) / 1000.0,
              (GetTimeMicros() - t0) / 1000.0);

Becomes:

    Tic t0;

    // do some stuff

    LogPrintf("Stuff took %s msec\n", t0.msecStr());

    Tic t1;
    // do more stuff

    LogPrintf("More stuff took %s msec, %s msec total time elapsed\n",
              t1.msecStr(), t0.msecStr());

I believe the baove is easier on the brain and takes less cognitive load
to accomplish.  It's as easy as `Tic` and then `msecStr()` later!

This code comes from Fulcrum, and I use it there all the time.  Note
that Tic only is 8 bytes so its overhead is identical to just doing it
all manually.

Test Plan
---

A unit test was added to ensure the sanity of this class and of the
underlying `GetProfTimeNanos()`.

- `ninja check`

Advanced test: Also build this for Windows (using either Gitian or
depends based) and run test_bitcoin.exe on that to run the check there,
if you like. I already did this here and it's green.

Note that on Windows it uses native calls to get the high precision time
from Windows itself -- which provides a very high quality clock.  Sadly,
Windows C++ libs based on mingw have been known in the past to provide a
very poor clock with low precision, which is why we opted to use
Windows's own native calls here to get sub-microsecond precision.