C++17: Replace boost::optional with std::optional
Summary
---

Now that we have switched over to C++17, we can remove 1 more boost
dependency: boost::optional.  This MR is done in the interests of code
clairty and simplicity.  As we move forward people would be using
std::optional anyway -- and having *two* different optional types in the
codebase would just be messy, so we need to do this sooner or later.
Might as well do it now!

The C++17 version is nearly identical API-wise (it's a subset of boost's
API, which offered a few extra overloads and convenience methods that
are non-essential).  The port over to using std::optional was a breeze,
as a result, since the semantics and everything are identical.  In fact
I believe the C++ standards committee used the boost implementation as
the basis for what they did (reducing it down to the bare essentials and
removing redundancies).

I carefully went over the entire codebase and removed it.  Note that we
already had "optional.h" which was intended to be a temporary file that
provided "C++17-like" optional via boost, without actually requiring
users to use boost directly.  That has been removed from the codebase as
well since it is now not needed.

Test Plan
---

- `ninja all check-all`
- Optional (heh): Also build with autoconf