C++17: Replace NODISCARD with [[nodiscard]]
Summary
---

Now that we are on C++17, the awkward NODISCARD macro (which is not even
available on all platforms theoretically) has been replaced with the
guaranteed-available [[nodiscartd]] attribute.

I went through the entire codebase and did a search-and-replace. This
one was pretty straightforward.  This is a pure nit/refactor commit with
no changed behavior.

Test Plan
---

- `ninja all bench_bitcoin check-all`