[qa] Update the copyright year in the GUI to 2021
Summary
---

When testing another MR I noticed the copyright year as appears on the
splash screen and in the "About" box for Bitcoin-Qt is not updated to
2021.  This MR fixes that.

Note I also went ahead and updated the copyright year in comments for
a few files that I edited for this commit, but not for all files in
the codebase, as that would be a bit much.

Test Plan
---

- Run the GUI, note that the splash screen should now say 2009-2021. The
  about dialog should also say 2021.
- (Optional) `ninja all check-all`

I believe the autoconf/automake version already had the updated year,
just the cmake build did not.