Remove version info from Node window
Makes the following changes to the Information tab of the Node window:

* Removes Client version. This could already be found in the About
dialog.
* Moves User Agent from General section to Network section.
* Removes BerkeleyDB version (only shown if built with wallet support).
Instead, this information is now added to the About dialog, and is now
also available in the `--version` CLI option output.

Overall, this puts version information where you expect it, and frees up
screen real estate in the Node window.

Test plan:

* `src/qt/bitcoin-qt`, check Node window (Information tab) and About
dialog
* `src/bitcoind --version`
* `src/qt/bitcoin-qt --version`