Double Spend Proof (dsproof-beta) core code
Co-authored-by: Calin Culianu <calin.culianu@gmail.com>

The double spend proof code detects double spends, sends an INV to peers
and handles getdata requests by introducing a new 'dsproof-beta' message.

This adds a DoubleSpendProof and a DoubleSpendProofStorage set of classes
which are as you'd expect. The storage is owned by the mempool and the
mempool will delete DSPs from the DSPStorage as the transaction they
belong to go away (for instance if they are mined). The link between a
transaction in the mempool and its proof are by an addition to the
CTxMemPoolEntry class.

The dsproof related classes can be found within `src/dsproof/`.

Known limitation: the code currently only manages to create DSPs when
the UTXO it double spends is a p2pkh. This helps for the majority of the
cases on the net, but this issue is the main reason we have the name
'beta' in the message name.

For tests and documentation please refer to subsequent commits.

Test plan (for these core code changes only):
- build using the various ways (cmake and autoconf)
- run the unit and extended functional tests to verify nothing broke
  (all tests should still pass)