Refactor: Make SeedSpec6 inherit from CService (and ultimately CNetAddr)
This helper class really should have always inherited from the core net
type CService.  The fact that it didn't do so originally was a code
smell.

As a result of this refactor, lots of code could be deleted and things
were simplified.  No std::memcpy to convert from CService to CAddress
and other code smell were removed.