Remove C macro ARRAYLEN in favor of C++17's std::size
Summary
---

The sourcecode itself asks us to do this in `util/strencodings.h`:

    // Note: ARRAYLEN should be replaced by std::size once we switch to C++17 or above
    #define ARRAYLEN(array) (sizeof(array) / sizeof((array)[0]))

As such, in the interests of code quality and type safety, we removed
the C-style `ARRAYLEN` macro in favor of C++17's template-based
`std::size` which is safer.

Also in this commit: While we were at it we modernized/fixed the
assignemnt of `vFixedSeeds` in `chainparams.cpp` to use the much safer `std::begin` and
`std::end` to get to the beginning and end of the array, rather than the
pointer arithmetic that was being done before.

Test Plan
---

Compiling everything should be enough, but if you like also do:

`ninja all check-all`