Normalise address input on blur
Similar to amount input fields (`BitcoinAmountField`), this normalises
the contents of address input fields (`QValidatedLineEdit` with
`BitcoinAddressCheckValidator`) when they lose focus.

In default configuration (`-usecashaddr=1`), this implies:

* Legacy base58 is converted to CashAddr
* Uppercase CashAddr is converted to lowercase
* CashAddr prefix is inserted if omitted

Because the normalisation happens on blur, the send confirmation dialog
will also use the normalised/canonical address format.

This MR also fixes a bug in the paste button next to the address field
on the send tab: it no longer allows pasting invalid characters. It now
focuses on the input so you can see the pasted address in the original
format, and then see the address conversion/normalisation happen upon
blur.

UX considerations:

* Seeing the address conversion on-screen makes clear to the user that
different notations of an address are really the same address.
* By removing the possibility of having the BTC address format in the
confirmation dialog, ambiguity is avoided about whether you are making a
BTC transaction or a BCH transaction.
* Newbies entering a BTC address may freak out if they see a "completely
different" address in the confirmation dialog. They should freak out -
this reduces the risk of accidentally sending BCH to BTC addresses.
* You can now use our GUI as a makeshift address convertor without
actually sending transactions.

Technical changes:

* `QValidatedLineEdit` got a public `validate` method similar to
`BitcoinAmountField`, replacing the previous `checkValidity` method.
* The `checkValidator` of `QValidatedLineEdit` is now allowed to modify
the contents of the input field, consistent with the `QValidator`
interface.

Test plan: `ninja bitcoin-qt check` and play around with the GUI