Seeder: Fix getaddr interval to actually query address lists once a day
Summary
---

The seeder was intended to fetch address lists of nodes once a day:

https://github.com/sipa/bitcoin-seeder/commit/966d040f8d976bc824821c5b2bb778363b5128fa

However, `ourLastSuccess` is updated each time a node is successfully
connected, which can be up to every 15 minutes. This effectively means
that address lists are only fetched when a node has been down for a day,
which means the seeder only queries address lists from new or unreliable
nodes.

To fix this, we add the new timestamp `lastAddressRequest` to the
database, which means the format is now version 5. This timestamp is
used to keep track of when we last send a `getaddr` / `getheaders`
request to a node.

Test plan
---

* `ninja check-bitcoin-seeder`
* Enable logging and observe address list being queried daily