[backport] ZMQ: Small cleanups in the ZMQ code
Co-authored-by: Calin Culianu <calin.culianu@gmail.com>

Summary
---

This is a backport of [Core#13686](https://github.com/bitcoin/bitcoin/pull/13686).

While working on !1045 I noticed some un-idiomatic C++ usage and a
potential memory leak. I went and looked at the Core repo and it turns
out they have since updated the ZMQ subsystem to be better and they
fixed the leak.

Original commit message:

    This contains various small code cleanups that make the ZMQ code easier
    to read  and maintain (at least in my opinion). The only functional change
    is that a potential memory leak is fixed that would have occured when a
    notifier is removed from the notifiers list after its callback function
    returned false (which is likely not relevant in practice but still a bug).

Note that this depends on MR !1045 (it is a commit on top of it).

I adapted this code a little bit to alphabetize the headers and also
other minor tiny nits. It's 99% the Core commit, essentially.

Test Plan
---

- `ninja all check`
- `test/functional/test_runner.py interface_zmq`
- `test/functional/test_runner.py rpc_zmq`