Add seeder: bchseed.c3-soft.com
Summary
---

This is an optional MR, just in case we want more seeders that we
control.  It adds the `bchseed.c3-soft.com` seeder to the MainNet seeder
list.  I keep this server updated and I also will be sure to always run
the latest stable seeder we have (with all bugfixes).  The seeder itself
has been running since October.

The server is my sever hosted by a major provider for my software company,
c3-soft, and it has an uptime of ~1256 days as of the time of this writing:

    root@c3:/home/bchseed# uptime
     08:51:07 up 1256 days,  9:58,  4 users,  load average: 0.10, 0.12, 0.13

It's my hope that by adding this, it would help with testing too since I
am more likely to update the seeder to the latest master than some of the other
places that host a seeder in that list, and so we can catch issues earlier
(hopefully there will never be any!).

Test Plan
----

- `ninja all check check-bitcoin-seeder`
- From the CLI on a Linux box: `dig bchseed.c3-soft.com` <-- should
  return random different results each time.