GUI: Fix icon for About menu entry
This makes the icon of the About entry in the Help menu and the
Show/Hide entry in the systray menu recognisable as BCHN logo, instead
of just an oddly shaped black blob.

Based on Leandrodimarco's graphics. I used his wireframe logo variant
and exagerrated the wire thickness for proper display as small icon.

Test plan: check icon in Help menu