Block index: Add an additional sanity check when loading the block index
Summary
---

This fixes issue #244. The core of the problem encountered in the issue
was that rows for block headers were missing from leveldb (likely due
to some sort of corruption). The block index loading code assumed that
all headers referenced by other headers (`hashPrevBlock`) would be present.
If they are not, it would never detect this and just crash later (with an
assertion failure) when building the skiplist. See issue #244.

This MR adds a check after the block index is loaded to ensure
that ALL of the headers in the map are not "0 initted" (a 0-initted
header can only happen if a leveldb row was missing for that header).

Note that an invariant in this app is that headers that are referenced
as previous headers from *other* headers should never be missing from the
DB, (even in a weird pruning situation) so this check is safe.

This extra sanity check adds no more than a few milliseconds to the
loader process for 1.5 million headers on my setup here, and it at least
makes the app produce a sensible error message if the sanity check fails,
rather than a crash as seen in #244.

With this sanity check in place, in bitcoin-qt it actually leads to a dialog
box asking you if you want to reindex!  (Which is better than what we had
before: a crash).

Test Plan
---

You are going to need to reproduce the situation encountered in issue #244.

To reproduce the issue, try this against master:

1. Set up a datadir, ideally synched (but it doesn't have to be to
   reproduce the crash itself).
2. Copy the `index` folder from the tarball linked in issue #244,
   replacing your datadir/block/index folder completely with it (be sure
   to rename your index folder first or to delete it to get it out of
   the way!).
3. Start bitcoind or bitcoin-qt and observe it crash with an assertion
   failure.

*Alternate method*: If the above is too complicated -- you can also reproduce
the crash by randomly deleting a row in the leveldb blocks/index database.
You will need to use an external tool to do this and you can pick a random
blockindex row to delete (all rows starting with the byte `'b'` are block index
rows).  Google for leveldb-cli to find such a tool.

Next, after having set up a corrupt block index using one of the above methods,
and after having reproduced the crash seen in #244, try loading the app using
this commit -- it shold now give you a more intelligent error message telling
you what to do next to fix ths situation.

**Also to test:**

- Run a (non-corrupted) pruned node as well with this commit, ensure it
  loads ok.
- Run a (non-corrupted) full node with this commit, ensuring it loads
  ok.
- Start up a brand new blockchain synch with this commit, aborting it in
  various places, restarting the app and then starting it again,
  ensuring nothing broke.
- Start up a new pruned node blockchain synch as above, killing it and
  restarting it and ensuring it still loads each time.