Code quality: Fix encapsulation of CTxMemPoolEntry::dspId
Summary
---

While studying the current mempool code I noticed a code quality issue
introduced by me (sorry!).  99% of class CTxMemPoolEntry is well
encapsulated with private data fields and public getters and setters --
except for the 'dspId' that I added to the class. Since the mempool code
already is mostly unencapsulated properly (sigh!) except for this class,
which happens to be well encapsulated, I figured I'd not make things
worse but make them better.

This commit is just a code quality commit that hides 'dspId' as private,
and adds getters and setters for this field.  It's a nit -- but at least
it means CTxMemPool stays as encapsulated as it always was even after
adding dsproof, rather than losing some of that.

Test Plan
---

This is entirely a code quality / nit commit -- no behavioral changes to
the codebase are introducted. However:

- `ninja all bench_bitcoin check`
- `test/functional/test_runner.py bchn-feature-doublespend-proof`
- `test/functional/test_runner.py bchn-rpc-dsproof`
- `test/functional/test_runner.py interface_zmq`