Refactor the SipHash-based salted hashers to all use common code
Summary
---

While researching the mempool modifications, I noticed that a lot of the
places in the codebase have a common need that they reproduce locally:
namely, creating a salted hasher for an unordered_map.  Various parts of
the codebase seem to duplicate this effort: txmempool, dsproof storage,
coins, netaddress, and others. It's all just boilerplate that does the
same thing.

So, in the interests of maintainability and code quality: I have
refactored all of this to use the same common code and common classes.

Summary of changes:

- Minor "bug" fix: The salted hashers before were using randrange which
  returns values always < UINT64_MAX -- added a `GetRand64()` to random.cpp
  that actually returns a random 64-bit covering the full range.
- Added a new: `util/saltedhashers.cpp` & `.h`, which contains the common
  boilerplate code for setting up a salted hasher, and provides a few of
  the more common-used onces in 1 place (hashers for uint256, txid,
  coutpoints, and bytevector).
- The bytevectorhash files were removed and instead folded into the
  above files.
- The places that set up their own salted hashers have been changed to
  call into this common code (such as txmempool, netaddress code, and
  other places).
- Created a utility class in hash.h for hashing any serializable with
  SipHash (used by netaddress code -- before it was doing this anyway
  privately -- it's probably better to export this facility for other code
  later that may want make use of it).
- I noticed a dependency problem in the CMake and makefile where
  hash.cpp was being used but not included in the "common" lib. It was
  being included in libbitcoinconsensus lib a step above but this is not
  100% kosher so it has been included twice so that the common lib remains
  self-consistent as a lib (many other files are included twice such as
  `util/strencodings.cpp` -- this is not that unusual and the link
  editor just removes dupes; also cmake doesn't compile them twice).
- Made the code in crypto/siphash.h `noexcept` since it can never throw
  (it operates on everything in-place).

Test Plan
---

While it would be nice if there were unit tests for the individual
hashers, none exist. This is a pure refactor and no behavioral changes
are introduced.  If existing code continues to function, then we are
good (there is code built on top of these hashers that is unit tested and
it would fail if this change were to break anything).

At any rate:

- `ninja all check-all`

Also do a build with the autotools `--deprecated-build-system` and ensure
it still builds there as well.