bitcoin-qt: Remove ifdef i386 stuff in the version string builder
Summary
---

This is a code quality commit.  The #ifdef stuff for `__i386__` versus
`__x86_64__` is very x86-specific. It's also kind of not idiomatic C++
anymore to include such ifdefs in UI code.

Also the comment is wrong -- it's not "ambigious" what platform bit
width you are on, ever. You check the size of a pointer.  If it's
64-bits, you are 64-bit. If not you are not.  There is no ambiguity here
at all.

This commit does introduce a "change" in that now the version string to
bitcoin-qt includes the build's arch width.

Test Plan
---

1. Linux: `bitcoin-qt -version` <-- check that 64-bit now appears in the
   version info.
2. OSX: `bitcoin-qt -version` <--- 64-bit now appears in the version
   info
3. Windows: `bitcoin-qt -version` <-- no change from previous