Readability: use proper algorithms.
## Summary

Avoid for loops when a known (and already named) algorithm can be used.
This improves readability, prevents errors, and in some cases improves performance.

## Note

It was designed refactor `onlyUnconfirmed()` function in `miner.cpp`
but it was recently deleted in !1078.
It can still be used on many other places, such as:

https://gitlab.com/bitcoin-cash-node/bitcoin-cash-node/-/merge_requests/1114/diffs#6409d86f7f9555776459d15b1ee46856cc8732e8_165_168

## Test Plan

* `ninja check check-functional`