mempool & dsproof: Make orphans stickier, make tx removal "orphanize" dsproofs first before removal later
Summary
-------

When working on the mempool code, I noticed that dsproofs are deleted
when a tx is confirmed in a block.  Later on, if there is a reorg, the
proof is now gone and cannot be reapplied to its tx even if the tx is
resurrected and put back into the mempool.  It may be beneficial for proofs
for confirmed tx's to actually stick around for a while so that they can be
reapplied should a mempool tx become "resurrected".

While not absolutely critical -- it is nice for known dsproofs to be
"sticky" to their tx's for a time after a tx is confirmed in a block.

This MR basically ensures that dsproofs aren't so quickly forgotten.
It closes #263 .

What's changed:

- `CTxMemPool::clear()` has a new optional boolean flag to not clear orphans
- `DoubleSpendProofStorage::clear()` also has a similar flag
   - Added a unit test for this.
- `DisconnectBlockTransactions::importMempool()` (used during reorg) now leaves orphans intact.
- When a tx that had a dsproof confirms (or is removed from mempool for any reason), we instead merely "orphanize" its dsproof.
   - This means that rather than instantly deleting the associated dsproof, we extend its lifetime a bit and instead put the dsproof into the orphan pool.
- If there is a reorg later, the dsproof will be a "rescued orphan" and reapplied to the tx.
  - The normal code path for rescued orphans will fire, including a ZMQ notification and a re-broadcast of the dsproof INV to peers.
    - Peers that have the proofs already will ignore these INVs
      (in fact, they may never even get the INVs due to the way the INV bloom filtering works in Bitcoin).
    - Peers that do not will be happy to hear about a proof they may have missed in the past.
- If there is no reorg after 90 seconds, the "orphanized" proof will be automatically cleaned by the dsproof storage orphan cleaner task that runs periodically.
  (This is the normal lifecycle of any orphan).

Functional tests have been updated to test the new behavior.

Test Plan
---------

- `ninja all check check-functional`