update ChainTxData for main, test3, and test4 networks
Summary
---

This data is not used for anything critical other than the Bitcoin-Qt UI
to display estimated progress info, I believe.

It has been updated for the 3 networks since we will do a release soon.

The tx rate estimate was calculated for test3 and test4 by just taking the
average rate since the last time we did this calculation.

For mainnet the rate went up significantly since last time, so I simply used
the last week's avg tps rate to arrive at that number, as published on
https://fork.lol .

Test Plan
---

- `ninja all check`
- Verify that the changed block height and timestamps match.