Update the chainparams update scripts to NOT touch scalenet chainwork
Summary
---

In recent work in MR !1173, we accidentally almost updated the scalenet
chain params.  Since scalenet is intended to reorg back to block 10,000
in the future, it is not a good idea to update the chain params for this
network since it would make reorging problematic.

This MR adds a sys.exit() call to the `make_chainparams.py` script if it
encounters scalenet.  A dev will need to manually edit that script if
they wish to actually update scalenet.

This extra safety feature avoids the situation where someone may
inadvertently update scalenet's chain params, commit the change to
master, then publish a release with the updated chain params, making
reorgs of scalenet troublesome.

This MR closes issue #293

Test Plan
---

1. Review
2. Try it out:

- `cd devtools/chainparams`

- Run `bitcoind` on any network **OTHER** than scalenet.
- `./make_chainparams.py -a host:port -c /path/to/notscalenet.conf` <--
  this should output some chain params (2 lines of hex)

- Run `bitcoind` **OON** scalenet.
- `./make_chainparams.py -a host:port -c /path/to/scalenet.conf` <--
  this should output an error message