bench: Add removeForBlock benchmark
Summary
---

This MR is part of on-going work to optimize
`CTxMemPool::removeForBlock`, as discussed in issue #161.

As such, it adds a benchmark for this function versus a full mempool of
~450k unchained txns.

The bench calls removeForBlock repeatedly with block txns from a ~32MB
block versus a very full mempool.

Also an 8MB block version is provided as well.

Test Plan
---

- `ninja all check bench_bitcoin`
- `src/bench/bench_bitcoin -filter='RemoveForBlock.*'`