Fix for rare failure in the bchn-rpc-getblocktemplatelight functional test
Summary
---

This closes issue #300.

Explanation for the failure seen in #300:

It's possible for this functional test to (falsely) fail spuriously due
to the "cached block template" mechanism interfering with the test.  To
ensure that it does not, the fix was to call `self.wait_for_txs` for each
of our two nodes (not just for the first one), which waits until a new
block template is "ready" and until the old one expires (5s max usually).

With this fix the test should no longer spurously fail.

Test Plan
---

- `ninja all && test/functional/test_runner bchn-rpc-getblocktemplatelight`

If you can reliably reproduce the failure,  then by all means try this
commit and try master and see this commit fixes it.

Due to the timing nature of the failure, it's hard to reproduce. I was
unable to do so here on my setup, but I am sure of the underlying issue,
having studied the code again with a critical eye after seeing #300's
error output.