Disable test: bchn-rpc-getblocktemplate-timing.py
Summary
---

This test is extremely unreliable in the CI environment due to the CI
env being "noisier" timing/CPU-wise (it likely runs in a VPS).

As such, due to the relatively high false failure rate of this test on
CI, we are better off just disabling it.

A directory was created inside the functional/ test directory called
"disabled".  We can move tests into here to keep the code around in case
we want to fix them, but otherwise disable the tests from ever normally
executing.

Test Plan
---

- Observe CI didn't run this `bchn-rpc-getblocktemplate-timing` test.