Fix `-txbroadcastinterval=0` leading to no txs sent + add functional test
Summary
---

The -txbroadcastinterval=0 argument should mean
"send txs as fast as possible, without an inv delay".
For values of -txbroadcastinterval=1 or 2, 4, 100, 500, etc
the delay keeps increasing, but for 0 it ends up never sending tx invs!

I am pretty sure this is not the intended design.

This MR fixes the issue (explained in #309).

It also adds a functional test to detect if this issue should ever regress.

Test Plan
---

- On this branch, do: `ninja all && test/functional/test_runner.py bchn-feature-txbroadcastinterval`
  And observe the test runs ok.
- Copy the bchn-feature-txbroadcastinterval.py test to a tmp dir, switch to master,
  copy the test back into ../test/functional/.
  And run the test: `test/functional/test_runner.py bchn-feature-txbroadcastinterval`

You should observe it fail off master but run ok off this MR.