Save dsproofs across restarts to datadir/dsproofs.dat
Summary
---

This MR adds dsproof saving to disk on node shutdown, and reload
from disk on node startup.

Motivation: While I was doing some testing and developing I noticed that
dsproofs get lost across restarts.  Sometimes node admins need to bounce
nodes and it would be nice if along with the mempool being saved/reloaded,
the dsproof state were also preserved across restarts.

This will become increasingly important in the future when wallets that
connect to servers such as Fulcrum start depending on seeing and
querying the dsproof state of particular transactions.

Description of Change
---

The dsproofs get saved to `datadir/dsproofs.dat` in a mechanism similar to
how the mempool is persisted to `datadir/mempool.dat`.  This is done once on
node shutdown.

On node restart, this file is loaded before the mempool is loaded and the
proofs are put into dsproof storage as "orphans".

Should the associated tx for the proof show up in the mempool later when
the mempool is loaded, the proofs will be re-associated with their
transactions.

Should the associated tx not ever show up or should it get immediately
confirmed by new blocks (after a longer period of being offline), the proofs
are never used and will remain as orphans, to be reaped within 60 seconds
(normal orphan reaping timeout).

The existing CLI arg `-persistmempool` is consulted when deciding whether
to save the `dsproofs.dat` file (alongside `mempool.dat`).  This is because
the two files really go together data-wise (dsproofs are just metadata for
transactions in the mempool, after all).

Test Plan
---

The `bchn-feature-doublespend-proof.py` test was updated to test this new
mechanism.

- `ninja all check && test/functional/test-runner.py bchn-feature-doublespend-proof`
- Also, run the above test with `--nocleanup` and look at the `debug.log`
  for `node0`, to observe the new log messages about saving/loading dsproofs
  from disk.