update ChainTxData for main, test3, and test4 networks
Summary
---

This data is not used for anything critical other than the Bitcoin-Qt UI
to display estimated progress info, and also various RPCs that provide information only.

It has been updated for the 3 networks since we will do a release soon.

The tx rate estimate was calculated for by just taking the
average rate since the last time we did this calculation.

Test Plan
---

- `ninja all check`
- Verify that the changed block height and timestamps match.