[gui] Fix potential unaligned access in bitcoin-qt image manipulation code
Summary
---

When compiling for ARM32, I noticed some potential unaligned access
warnings from the compiler.  The code in question was
reinterpret_casting char arrays to uint32, which can be problematic on
some ARM cpus that do not properly support unaligned access.

Internally I'm pretty sure the Qt libs align the image data, so perhaps
this MR is just a paranoia MR.  But still, it's good to not do UB and/or
potentially unaligned access if we can avoid it.

This MR uses the idiom of copying data using std::memcpy to an aligned
buffer to access it (and copying it back).

Test Plan
---

- `ninja all check`
- Run the Qt GUI on one of the testnets, verify that the splash screen
  image has a modified color (it should be Green for testnet3, purple
  for testnet4, etc).
- Run the Qt GUI on Linux on one of the testnets. Try and send yourself
  some funds, see that the DBUS notification image looks ok.