Fix fuzzer to compile ok on some newer systems
Summary
---

While testing !1208 with a fuzzer, I observed that it fails to compile
on newer Ubuntu 21.10 which uses clang-13.

This is due to a missing `#include <string>` in `fuzz.h`.  On older
systems apparently the header is "accidentally" included by other libc++
headers, but on newer systems it is not, thus we get a compile error due
to unknown type `std::string`.

Note that the error is not clang-13 itself, but rather the header environment
installed on the system, thus on some clang-13 systems this error does not occur.

Closes issue #341.

Test Plan
---

Since this is a trivial commit, I would say this test plan is 100% optional.

But if you like, to reproduce the issue, do the following against both master
and this commit, using Ubuntu 21.10.

```bash
rm -rf build_fuzz && mkdir build_fuzz && cd build_fuzz
cmake -GNinja .. -DENABLE_MAN=OFF -DCCACHE=OFF -DENABLE_SANITIZERS="address;fuzzer" -DCMAKE_C_COMPILER=clang-13 -DCMAKE_CXX_COMPILER=clang++-13
ninja bitcoin-fuzzers link-fuzz-test_runner.py
```

Observe it fail to compile on master but succeed with this commit.