Fix rpc_tests unit test to not have rare spurious failures
Summary
---

See issue #345.  Basically, this unit test was calling BOOST_CHECK()
from a thread, which is not supported by Boost (not thread safe).
This worked "most of the time" and led to the test mostly succeeding,
but with rare 0.3% spurious failure probability, as reported by a
project collaborator.

The test has been modified to not use BOOST_CHECK in a subthread in the
multithreaded case.

Closes #345.

Test Plan
---

- Try to reproduce the issue seen in #345 by e.g. doing this from
  master:

    $ let i=0; while ((++i < 1000)) && src/test/test_bitcoin -t rpc_tests; do nop=1; done

- If you do manage to reproduce the issue (note, sometimes it never
  happens -- on my macOS system I cannot reproduce), try this branch and
  see if it fixes the issue.