Fix testlib_tests failing when compiling with ASAN and/or TSAN
It turns out that the BCHN_CHECK_ASSERT facility (which uses some
unorthodox methods to detect assert() firing) is sufficiently
unorthodox in that ASAN and/or TSAN interfere with it.

As such, we disable these checks if the binary was compiled with either
of these two sanitizers enabled.

This fixes the unit tests failing when testing ASAN builds.