[serialization] fix GCC variable shadowing warning
Summary
---

The new `SER_READ` and `SER_WRITE` macros introduce a lot of variable
shadowing (`-Wshadow`) warnings. To fix this we use the lambda default
capture to get the needed variables into the lambda and use an
`if constexpr` to conditionally compile the lambda only when needed.

This also changes `MAX_SIZE` and `MAX_VECTOR_ALLOCATE` to constexpr.

Test plan
---

* `ninja all check-all`
* check that no more `-Wshadow` warnings show up