Remove dependency on git-lfs; generate .cpp data files as part of build
Co-authored-by: Andrew#128 <andrew-128-github@protonmail.com>

This MR closes #333.

Summary
-------

This commit eliminates the need for git-lfs. Instead, the .raw files
have been compressed with bzip2 -9, and have been added to this
repository directly as .raw.bz2. They are about 40% smaller compressed than
uncompressed.
The build script convert-raw-files.py has been modified to know how to
uncompress the data before generating C++ code for embedding.  The data
is now encoded as a base64 string in the C++, and is decoded at most
once the first time the data is read by client code. The base64 encoding
is much more efficient than the wasteful \XXX octal encoding that was
used before. It's about ~3 times more compact than previous, which should
lead to faster compilation times for bench_bitcoin.
A build step was added for the bench_bitcoin target which runs the
script to generate the data .cpp files (which get compiled into
bench_bitcoin as static data). The .cpp files are sent to the
"build" directory (thus they don't live in-tree).
The above change adds an additional implicit dependency for one wishing to
build bench_bitcoin: Python3 must be present on the system. However,
considering that other "dev-oriented" targets require python (such as
check and check-functional), it should not be a problem.
Within bench_bitcoin's benches, the API for accessing the data via
data.h has been slightly modified, in that now the external inferface is
a function returning a std::vector<uint8_t> &. This allows us to not
waste cycles and memory decoding or otherwise doing anything to the data
until it's actually needed.
The binary size of bench_bitcoin has been reduced significantly due to
the use of base64 encoding of the binary data (rather than the
inefficient \XXX octal encoding used previously).

Test Plan
---------

- `ninja clean`: Verify that there are no .cpp files in your build/src/bench/data folder.
- `ninja all`: Verify that there are no .cpp files in your build/src/bench/data folder.
- `ninja bench_bitcoin`: Verify that there are new .cpp files in your build/src/bench/data folder.
- `ninja clean`: Verify that the .cpp files seen above are now gone.

Also, this should probably be tested under various build conditions,
such as Gitian, depends, etc.