Fix walletpassphrase, walletpassphrasechange and walletlock RPC doc
These wallet RPC calls were not returning help at all or not returning it if the wallet
was not encrypted. And therefore not showing up in the generated documentation:

* walletpassphrase
* walletpassphrasechange
* walletlock

A conditional checks on `IsCrypted()` was also removed in `encryptwallet` RPC,
so that the help shows up even if the wallet is not encrypted.
Some dead code was removed in these RPC calls.

Test Plan
---------

- `ninja all check-all`
- `ninja doc-html` and browse the generated documentation (in `doc/html/`)
   to check that the pages for the above are included in
   `API documentation` -> `JSON-RPC interface` -> `Wallet` section.
   Check that the generated documentation looks OK.

Closes #378