Add internal string split function
The motivation is eventual Boost elimination (#390).

The boost string split algorithm is used in a number of places throughout
the code.

This MR creates an internal string split function, along with other string
utility functions that already exist.  It currently just wraps the boost
split algorithm, in order to ensure consistency.

Tests are added that verify this functionality. This paves the way for a
future merge request which will replace all instances of boost split in
the code with the internal function, in a way that is tested and known safe.
Subsequently, the function can be reimplemented locally to avoid using boost,
with the confidence of established tests available to verify any new algorithm
and identify regressions.

Test plan
---------

- `ninja check-all`
- `src/test/test_bitcoin --log_level=all --run_test=util_tests/util_Split -- DEBUG_LOG_OUT | fgrep -av 'is skipped because disabled'`