Add logging category `httptrace` to log all RPC requests/responses
Co-authored-by: Axel Gembe derago@gmail.com

Summary
-------

This affects the RPC server (-server) as well as the REST server (-rest).
If -debug=httptrace is supplied, the log file will contain all http requests
and responses transferred to/from the HTTP server.

This is a low-level debugging option created to address issue #387.
It closes #387.

Because the log could potentially fill up very quickly, and especially
since it may contain sensitive information, this -debug option is
not enabled by default if supplying -debug=1 or -debug or
-debug=all (as is done by the tests).  It ONLY is enabled if explicitly
requested with -debug=httptrace.

Also in this commit (cleaned up some things):

- Make use of NRVO for HttpRequest::ReadBody()
- Use C++17 std::optional for HttpRequest::GetHeader() rather than the less
  ergonomic (bool,string) pair.
  - Also GetHeader() was copying objects excessively when returning a pair. This has been addressed.

Test Plan
---------

- `ninja all check-all` or `ninja all check && test/functional/test_runner.py bchn-feature-httptrace`
- bitcoind -hh | less and check the new/changed help text for the -debug= option.
- Also start the node with -debug=httptrace and verify that it works and that the output it produces is to your liking.
  To see the HTTP trace you should hit the RPC server with some RPC requests (using bitcoin-cli).
- Try the REST server (-rest) as well and check that trace output is generated for it as well.
- Try the above without -debug=httptrace and verify that logging of the trace is disabled, even if -debug=all is specified.