Run tests with eatmydata if it's available
A small addition to the build system: detect if the eatmydata program is
available in the path, and configure the build-launched tests to run with it.
Enabled by default, it can be disabled with cmake -DENABLE_EATMYDATA=OFF.
The option is also documented in cmake itself, and mentioned to the user
when eatmydata is detected.
This change only affects tests launched from the build system (e.g. with
`ninja check-all` - tests launched manually are unaffected.

For more info on motivation, see

  https://gitlab.com/bitcoin-cash-node/bitcoin-cash-node/-/issues/380#note_794191199
  
This change also replaces hard tabs with spaces in CMakeLists.txt;
I recommend hiding whitespace changes when reviewing.

Associated pull request to add eatmydata to the gitlab CI docker image:

  https://github.com/dagurval/docker-bcn-buildenv/pull/7

Test plan
---------

Run and time main test suite, before and after:

- `time ninja check-all`