Replace `boost::split` with a custom implementation in `util/string.h`
Summary
-------

This is part of on-going work to reduce our dependence on boost headers
and the boost library in general, as discussed in issue #390.

The new implementation of Split is very much compatible with the
boost::split implementation we used before, however its genericity has
been reduced to only operate on std::string_view (which itself can be
implicitly converted for zero-cost from std::string arguments).

The previous boost::split technically could operate on any container
with any value_type.  In this codebase Split is only used to operate
on std::strings, so narrowing it down to what we need is probably a good
thing.

Side note: Interestingly, the boost version had problems with std::string_view,
whereas this version does also accept std::string_view as its input string
(in fact, that's what it uses as the preferred argument type). Boost's
incompatibility with std::string_view was likely due  to string_view
lacking an (InputIterator begin, InputIterator end) constructor in C++17,
which was  added in C++20. So in some ways, it can be argued that this
version is more useful for actually operating on strings.

Test Plan
---------

- `ninja all bench_bitcoin check-all`
- Try the benchmarks: `src/bench/bench_bitcoin -filter='String_Split.*'`.
- You can compare the above benchmark versus boost by switching to master
  prior to this commit

In some cases this is faster, in some cases boost is faster, but overall
there is no great difference in performance on my system.