Replace boost::replace_all with internal ReplaceAll and std::string functions
Part of #390 - Boost elimination.

This MR eliminates another Boost dependency - boost string replace algorithm.
It introduces an equivalent ReplaceAll function which acts as a drop-in
replacement, and adds associated tests for the function which are equivalent
to the unit tests Boost itself uses for boost::replace_all and boost::erase_all.
In one place boost::replace_all was simply used to strip leading OP_ substrings,
so that has been simplified as a conditional erase.

This MR also eliminates some unused string algorithm includes.
In conjunction with !1382, this allows elimination of all reliance on the boost
string algorithm library.

Test plan
---------

- `ninja check-bitcoin`
- `ninja check-functional`