[build] Suppress ABI change build warnings when using GCC >= 7 on ARM
Summary
---

There was a bug fixed in GCC >= 7 that results in an ABI change and the
GCC developers have decided to output a warning in that case:

https://gcc.gnu.org/gcc-7/changes.html

We do not mix GCC versions so we should not be affected by this change
but the warning is making the ARM build very noisy. We suppress the
warning when we detect an ARM build with GCC version >= 7.

Test plan
---

* On multiarch Debian or Ubuntu, install `gcc-arm-linux-gnueabihf`,
  `g++-arm-linux-gnueabihf`, `qemu-user`, `libc6:armhf` and
  `libstdc++6:armhf`
* `cd depends && make build-linux-arm ; cd ..`
* `cmake -GNinja .. -DCMAKE_TOOLCHAIN_FILE=../cmake/platforms/LinuxARM.cmake -DCMAKE_CROSSCOMPILING_EMULATOR=qemu-arm-static`
* Check that there are no more `note: parameter passing for argument of type T changed in GCC 7.1` warnings