[test] Find emulator on path if it does not exist
Summary
---

Right now the test runner fails when `CMAKE_CROSSCOMPILING_EMULATOR` is
specified without the full absolute or relative path of the emulator.

The CMake `add_test_runner` function already starts the emulator from
the path if it can, but the python test runner doesn't.

This makes the test runner resolve the path from `config.ini` using
`shutil.which` when it can not find the file.

Test plan
---

* Do a cross compile for a foreign architecture:
* On multiarch Debian or Ubuntu, install `gcc-arm-linux-gnueabihf`,
  `g++-arm-linux-gnueabihf`, `qemu-user`, `libc6:armhf` and
  `libstdc++6:armhf`
* `cd depends && make build-linux-arm ; cd ..`
* `cmake -GNinja .. -DCMAKE_TOOLCHAIN_FILE=../cmake/platforms/LinuxARM.cmake -DCMAKE_CROSSCOMPILING_EMULATOR=qemu-arm-static`
* `ninja all check-all`
* This should pass the test suite.