Add a follow-up note assumptions made in CAddress w.r.t. extversion
As was observed in review of MR 1473, serialization code from Core in
protocol.h for CAddress makes assumptions that may not always hold for
BCHN, which has extversion support.

Extversion can send any data objects, theoretically, as part of the
EXTVERSION message.  If we were to ever decide to also send CAddress
instances using extversion (which it could support, in theory), then
assumptions made in the serializer for CAddress in protocol.h would no
longer hold.

This MR adds a note about that, so as to be defensive.