seeder: Use poll() instead of select(), if supported
This is a follow-up to !1386 in which the base net layer uses poll()
over select() on certain platforms (currently just Linux).

However, the seeder is still using select().  select() suffers from some
design problems in that if it happens to get given an FD > 1024 on
Linux, it won't work.

This commit takes the USE_POLL define, from compat.h, and if that is
defined, and <poll.h> is also present at compile-time, then poll() will
be used instead of select().