Fix to GetRandomDuration
The existing implementation from Core abused std::common_type in a
strange way, making it not very readable or obvious.  The decision to do
it this way made the code not very self-documenting, as evidenced by the
very long comment required to explain what is going on.

This change makes the code self-documenting and readable.