Thread-safety fix to address cache in CConnman
Since CConnman is a global object (acessed via g_connman), and since it
potentially can, in future updates to the code, have its
GetAddresses(Network) method accessed by any thread, it is prudent to
make the code thread-safe.  It is not currently thread-safe, but this
commit makes it thread-safe.

Currently the address cache used is not guarded by any locks.  This
happens to work ok in the current codebase because only the `msghand`
thread touches this object currently.  However, it is a potential
landmine to leave the code as-is, without locks, because
GetAddresses(Network) is a public method and it makes absolutely no
mention of the fact that *only* the `msghand` thread is expected to
call this.

This commit does the following:

- Guards CConnman::m_addr_response_caches with a lock.
- Also, a slight fix: Does not cache empty results, since doing so is
  not helpful towards privacy, and only serves to slow down address
  propagation without real benefit.