[doc] Update relaese-notes to talk about changes to `-bind=`
Summary
---

After !1496 was merged, it was observed that default behavior of
BCHN slightly changed with respect to which ports it binds to by
default. It now binds to an additional port on localhost only
for the tor hidden service if using -listenonion (enabled by
default). We needed to clarify this in the release-nodes. We also
needed to mention the new `=onion` syntax for the `-bind=` option.

Test Plan
---

Review wording, etc.