[qa] Bump version to 24.1.0
Summary
---

This is in preparation for the upcoming release.  It was decided in
Slack by maintainers that a minor version bump was warranted on account
of this version containing a significant update to a major subsystem:
the network subsystem was updated to support torv3.

Test Plan
---

`ninja clean && ninja && src/bitcoind --version` (verify version is updated)