Hardcode height for May 15, 2022 consensus upgrade
This allows us to avoid the expensive MTP check for the May 15, 2022 upgrade
and instead use pindexPrev->nHeight directly which is much faster.
We did this for phonon and other previous upgrades (but not Axion
for "reasons").

Note that we hard-coded the activation height for scalenet to 10,006
since we will reorg this chain anyway in the future bqck to height 10,000.
We chose 10,006 because other nodes may be relying on MTP time to activate
upgrade8 still, and we anticipate that after a reorg height 10,007 will
be the first block to accept the new rules (the activation height is always
1 less than the first fork block for new rules).

It's safe to retroactively do this to scalenet because upgrade8 actually
was a "reverse-soft-fork", or a loosening of consensus rules. So the scalenet
chain is valid using the new rules because it was valid under the older,
stricter rules. To keep things simple we simply retroactively pretend that
scalenet always allowed native introspection and other upgrade opcodes after
height 10,006.

We also removed the `-upgrade8activationtime` flag from BCHN.

This means we had to update 2 functional tests to remove the use of the
upgrade8activationtime flag (which simply tested for block rejection
pre-activation).

These tests were good to have but since the new rules are now active on
the real chain, and since we had to remove the activation time argument,
the two minor functional tests cases relying on this argument were removed.

Test Plan
---------

- Review the hard-coded heights are 1 before the fork block (activation is
  weird and we always hard-code the height that is 1-before the fork block!)
- `ninja all check-all`
- Optional: Run a full node on mainnet and sync it from scratch
  (can be done before release)
- Optional2: Run a full scalenet node off this MR and sync it from scratch
  to be paranoid