[OSX] Update to clang-8.0.0, MacOSX SDK 10.15 and set minimum sysver=10.14
This is required for the upgrade to Qt 5.15.3 to work on OSX.
It is a pre-requisite for MR !1523.

We upgrade the clang we use to build the OSX binaries from 6.0.1 -> 8.0.0,
thus bringing improvements to generated code and increased support for C++17.

Additionally, we were forced to update the OSX minimum system version from
10.12 (Sierra) to 10.14 (Mojave). This increases C++17 coverage and is also
required by the new Qt 5.15.3 in !1523. This means that it's possible that
the pre-built app we ship won't run on older macs, however at this point Mojave
is itself 4 versions back, so setting that as the minimum brings some benefits.
Namely.. dark mode will work once we upgrade to Qt 5.15.3, which is a huge
UI win for most users.

Test Plan
---------

- You need to update your downloaded MacOSX SDK from 10.14 -> 10.15
  (see instructions in commit's `doc/gitian-building.md` and `depends/README.md`)
- Add this to `gitian-builder/inputs`, and also untar it in `depends/SDKs`.
- Run gitian build for OSX as well as regular depends builds for OSX.
- Ensure gitian still works correctly, is reproducible, etc.