Remove install target definition for bench_bitcoin
Some versions of CMake (including the one used by the project's current CI)
produce the following warning:

  WARNING: Target "bench_bitcoin" has EXCLUDE_FROM_ALL set and will not be
  built by default but an install rule has been provided for it.
  CMake does not define behavior for this case.

Indeed it does not make sense for there to be an install definition for
an excluded target, and the existing definition attempts (incorrectly)
to define it as an exclusion anyway.

This small MR removes the install target definition for bench_bitcoin entirely.

Test plan
---------

- When running `cmake -GNinja ..` in a fresh build directory, verify that
  the above warning is not generated.

- Check the CI build logs in the latest pipeline for this MR to confirm likewise.