lint: Allow std::variant to appear in codebase.
All of the compilers and libraries we now target support this C++17
language feature. The last holdout was our macOS support, but since
we upgraded to 10.14 as the base deployment target in MR !1524,
std::variant is now supported on macOS as well.  As such, we can remove
this linting check from our codebase.